--- packages/consts/src/application-statuses.ts
@@ -1,6 +1,0 @@
-export const APPLICATION_STATUSES = [
-  "pending",
-  "accepted",
-  "waitlisted",
-  "checkedin",
-] as const;
