--- packages/consts/src/ethnicities.ts
@@ -1,12 +1,0 @@
-export const ETHNICITIES = [
-  "White",
-  "Black or African American",
-  "Hispanic or Latino",
-  "Asian",
-  "Native Hawaiian or Other Pacific Islander",
-  "Native American or Alaska Native",
-  "Middle Eastern",
-  "Two or more",
-  "Prefer not to say",
-  "Other",
-];
