--- packages/consts/src/genders.ts
@@ -1,7 +1,0 @@
-export const GENDERS = [
-  "Man",
-  "Woman",
-  "Non-binary",
-  "Prefer not to say",
-  "Other",
-];
