+++ packages/consts/src/index.ts
@@ -1,9 +1,2711 @@
-export * from "./majors";
-export * from "./schools";
-export * from "./shirt-sizes";
-export * from "./sponsor-tiers";
-export * from "./application-statuses";
-export * from "./countries";
-export * from "./states";
-export * from "./genders";
-export * from "./ethnicities";
+export const GENDERS = [
+  "Man",
+  "Woman",
+  "Non-binary",
+  "Prefer not to say",
+  "Other",
+] as const;
+
+export const MAJORS = [
+  "Computer Science",
+  "Computer Engineering",
+  "Electrical Engineering",
+  "Information Technology (IT)",
+  "Software Engineering",
+  "Aeronautics",
+  "Agricultural Business & Technology",
+  "Agriculture",
+  "Agricultural Economics",
+  "Agricultural Operations",
+  "Animal Sciences",
+  "Construction Engineering Technology",
+  "Construction Management",
+  "Fish and Wildlands Science and Management",
+  "Food Science",
+  "Animation",
+  "Art",
+  "Art History",
+  "Cinematography",
+  "Dance",
+  "Digital Arts",
+  "Drama",
+  "Fashion Design",
+  "Film",
+  "Game Design",
+  "Graphic Design",
+  "Industrial Design",
+  "Illustration",
+  "Interior Design",
+  "Multimedia",
+  "Music",
+  "Music Management",
+  "Music Performance",
+  "Photography",
+  "Studio Art",
+  "Visual and Performing Arts",
+  "Visual Design",
+  "Accounting",
+  "Actuarial Science",
+  "Advertising",
+  "Business",
+  "Business Administration",
+  "Business, Management, and Related Support Services",
+  "Econometrics",
+  "Economics",
+  "Entrepreneurship",
+  "Fashion Merchandising",
+  "Finance",
+  "Human Resources Management",
+  "International Business",
+  "Insurance",
+  "Management",
+  "Management Information Systems",
+  "Management Science",
+  "Managerial Economics",
+  "Marketing",
+  "Operations Management",
+  "Organizational Behavior",
+  "Real Estate",
+  "Sales",
+  "Supply Chain Management",
+  "Early Childhood Education",
+  "Education",
+  "Education (Other)",
+  "Elementary Education",
+  "Health and Physical Education",
+  "Language Arts Teacher Education",
+  "Mathematics Teacher Education",
+  "Middle School Education",
+  "Music Teacher Education",
+  "Physical Education Teaching and Coaching",
+  "Secondary Education",
+  "Special Education",
+  "Teacher Education",
+  "Athletic Training",
+  "Audiology",
+  "Clinical Laboratory Science",
+  "Cognitive Science",
+  "Communication Sciences",
+  "Community Health",
+  "Community Health Services",
+  "Dental Hygiene",
+  "Health and Wellness",
+  "Health Care Administration",
+  "Health Information",
+  "Health/Medical Preparatory Programs",
+  "Health Professions and Related Clinical Sciences",
+  "Health Services",
+  "Hospital and Health Care Facilities Administration",
+  "Medical Radiologic Technology",
+  "Nursing",
+  "Nursing Research",
+  "Nursing Science",
+  "Nutrition",
+  "Nutrition Sciences",
+  "Public Health",
+  "Public Health Education",
+  "Radiologic Technology",
+  "Respiratory Care Therapy",
+  "Speech-Language Pathology",
+  "Applied Communication",
+  "Biblical Studies",
+  "Communication",
+  "Communication and Media Studies",
+  "Creative Writing",
+  "Digital Communication and Multimedia",
+  "English",
+  "Foreign Languages and Literature",
+  "French Language and Literature",
+  "Gender Studies",
+  "General Studies",
+  "General Studies and Humanities",
+  "History",
+  "Humanities",
+  "Journalism",
+  "Liberal Arts",
+  "Linguistics",
+  "Mass Communication",
+  "Organizational Communication",
+  "Parks, Recreation and Leisure Facilities Management",
+  "Parks, Recreation and Leisure Studies",
+  "Philosophy",
+  "Public Relations",
+  "Public Relations/Image Management",
+  "Radio and Television",
+  "Religious Studies",
+  "Rhetoric and Composition",
+  "Spanish Language and Literature",
+  "Speech and Rhetoric",
+  "Talmudic Studies",
+  "Theology",
+  "Community Organization and Advocacy",
+  "Geography",
+  "International Relations",
+  "International Studies",
+  "Legal Assistant",
+  "Legal Studies",
+  "Political Science",
+  "Public Administration",
+  "Public Policy Analysis",
+  "Social Work",
+  "Sociology",
+  "Anthropology",
+  "Behavioral Sciences",
+  "Child Development",
+  "Criminology",
+  "Experimental Psychology",
+  "Human Development and Family Studies",
+  "Psychology",
+  "Aerospace Engineering",
+  "Applied Mathematics",
+  "Architecture",
+  "Astrophysics",
+  "Biochemistry",
+  "Biological and Biomedical Sciences",
+  "Biological and Physical Sciences",
+  "Biology",
+  "Biomedical Engineering",
+  "Biomedical Sciences",
+  "Cellular and Molecular Biology",
+  "Chemical Engineering",
+  "Chemistry",
+  "Civil Engineering",
+  "Computer and Information Sciences",
+  "Computer Systems Networking",
+  "Conservation",
+  "Earth Science",
+  "Engineering",
+  "Environmental Science",
+  "Environmental Studies",
+  "Exercise Physiology",
+  "Informatics",
+  "Information Science",
+  "Information Systems Security",
+  "Kinesiology",
+  "Marine Biology",
+  "Mathematics",
+  "Mechanical Engineering",
+  "Microbiology",
+  "Neuroscience",
+  "Petroleum Engineering",
+  "Physics",
+  "Physiology, Pathology, and Related Sciences, Other",
+  "Statistics",
+  "Web Design",
+  "Zoology",
+  "Apparel and Textiles",
+  "Corrections and Criminal Justice",
+  "Criminal Justice",
+  "Electromechanical Technology",
+  "Foods, Nutrition, and Wellness Studies",
+  "Forensic Science and Technology",
+  "Hospitality Administration",
+  "Hotel Management",
+  "Human Sciences (General)",
+  "Human Services",
+  "Law Enforcement Administration",
+  "Police Science",
+  "Sport and Fitness Administration",
+  "One Piece",
+  "Other",
+] as const;
+
+export const ETHNICITIES = [
+  "White",
+  "Black or African American",
+  "Hispanic or Latino",
+  "Asian",
+  "Native Hawaiian or Other Pacific Islander",
+  "Native American or Alaska Native",
+  "Middle Eastern",
+  "Two or more",
+  "Prefer not to say",
+  "Other",
+] as const;
+
+export const COUNTRIES = [
+  "United States",
+  "China",
+  "India",
+  "Germany",
+  "Russia",
+  "Afghanistan",
+  "Albania",
+  "Algeria",
+  "Andorra",
+  "Angola",
+  "Antigua and Barbuda",
+  "Argentina",
+  "Armenia",
+  "Australia",
+  "Austria",
+  "Azerbaijan",
+  "Bahamas",
+  "Bahrain",
+  "Bangladesh",
+  "Barbados",
+  "Belarus",
+  "Belgium",
+  "Belize",
+  "Benin",
+  "Bhutan",
+  "Bolivia",
+  "Bosnia and Herzegovina",
+  "Botswana",
+  "Brazil",
+  "Brunei",
+  "Bulgaria",
+  "Burkina Faso",
+  "Burundi",
+  "Cabo Verde",
+  "Cambodia",
+  "Cameroon",
+  "Canada",
+  "Central African Republic",
+  "Chad",
+  "Chile",
+  "Colombia",
+  "Comoros",
+  "Congo, Democratic Republic of the",
+  "Congo, Republic of the",
+  "Costa Rica",
+  "Côte d'Ivoire",
+  "Croatia",
+  "Cuba",
+  "Cyprus",
+  "Czechia",
+  "Denmark",
+  "Djibouti",
+  "Dominica",
+  "Dominican Republic",
+  "Ecuador",
+  "Egypt",
+  "El Salvador",
+  "Equatorial Guinea",
+  "Eritrea",
+  "Estonia",
+  "Eswatini",
+  "Ethiopia",
+  "Fiji",
+  "Finland",
+  "France",
+  "Gabon",
+  "Gambia",
+  "Georgia",
+  "Ghana",
+  "Greece",
+  "Grenada",
+  "Guatemala",
+  "Guinea",
+  "Guinea-Bissau",
+  "Guyana",
+  "Haiti",
+  "Honduras",
+  "Hungary",
+  "Iceland",
+  "Indonesia",
+  "Iran",
+  "Iraq",
+  "Ireland",
+  "Israel",
+  "Italy",
+  "Jamaica",
+  "Japan",
+  "Jordan",
+  "Kazakhstan",
+  "Kenya",
+  "Kiribati",
+  "Korea, North",
+  "Korea, South",
+  "Kosovo",
+  "Kuwait",
+  "Kyrgyzstan",
+  "Laos",
+  "Latvia",
+  "Lebanon",
+  "Lesotho",
+  "Liberia",
+  "Libya",
+  "Liechtenstein",
+  "Lithuania",
+  "Luxembourg",
+  "Madagascar",
+  "Malawi",
+  "Malaysia",
+  "Maldives",
+  "Mali",
+  "Malta",
+  "Marshall Islands",
+  "Mauritania",
+  "Mauritius",
+  "Mexico",
+  "Micronesia",
+  "Moldova",
+  "Monaco",
+  "Mongolia",
+  "Montenegro",
+  "Morocco",
+  "Mozambique",
+  "Myanmar",
+  "Namibia",
+  "Nauru",
+  "Nepal",
+  "Netherlands",
+  "New Zealand",
+  "Nicaragua",
+  "Niger",
+  "Nigeria",
+  "North Macedonia",
+  "Norway",
+  "Oman",
+  "Pakistan",
+  "Palau",
+  "Palestine",
+  "Panama",
+  "Papua New Guinea",
+  "Paraguay",
+  "Peru",
+  "Philippines",
+  "Poland",
+  "Portugal",
+  "Qatar",
+  "Romania",
+  "Rwanda",
+  "Saint Kitts and Nevis",
+  "Saint Lucia",
+  "Saint Vincent and the Grenadines",
+  "Samoa",
+  "San Marino",
+  "Sao Tome and Principe",
+  "Saudi Arabia",
+  "Senegal",
+  "Serbia",
+  "Seychelles",
+  "Sierra Leone",
+  "Singapore",
+  "Slovakia",
+  "Slovenia",
+  "Solomon Islands",
+  "Somalia",
+  "South Africa",
+  "South Sudan",
+  "Spain",
+  "Sri Lanka",
+  "Sudan",
+  "Suriname",
+  "Sweden",
+  "Switzerland",
+  "Syria",
+  "Taiwan",
+  "Tajikistan",
+  "Tanzania",
+  "Thailand",
+  "Timor-Leste",
+  "Togo",
+  "Tonga",
+  "Trinidad and Tobago",
+  "Tunisia",
+  "Turkey",
+  "Turkmenistan",
+  "Tuvalu",
+  "Uganda",
+  "Ukraine",
+  "United Arab Emirates",
+  "Uruguay",
+  "Uzbekistan",
+  "Vanuatu",
+  "Vatican City",
+  "Venezuela",
+  "Vietnam",
+  "Yemen",
+  "Zambia",
+  "Zimbabwe",
+] as const;
+
+export const APPLICATION_STATUSES = [
+  "pending",
+  "accepted",
+  "waitlisted",
+  "checkedin",
+] as const;
+
+export const SCHOOLS = [
+  "The University of Central Florida",
+  "Florida International University",
+  "The University of South Florida",
+  "The University of Florida",
+  "Florida State University",
+  "Florida Agricultural & Mechanical (A&M) University",
+  "Florida Atlantic University",
+  "Florida Gulf Coast University",
+  "Florida Institute Of Technology",
+  "Florida Polytechnic University",
+  "University of North Florida",
+  "University of West Florida",
+  "21st Century Cyber Charter School",
+  "ABES Engineering College",
+  "ACT Academy Cyber Charter School",
+  "AGH University of Science and Technology",
+  "AMC Engineering College",
+  "APPA Institute of Engineering and Technology",
+  "APS College of Engineering",
+  "ASPIRA Bilingual Cyber Charter School",
+  "ATME College of Engineering",
+  "Aalto University",
+  "Aarhus University",
+  "Abbey Park High School",
+  "Abbey Park Middle School",
+  "Abertay University",
+  "Abington Senior High School",
+  "Abraham Lincoln High School",
+  "Abraham Lincoln High School - Philadelphia",
+  "Academy at Palumbo",
+  "Academy of Technology",
+  "Acardia High School, Arizona",
+  "Achariya College of Engineering Technology",
+  "Acharya Institute of Technology",
+  "Acharya Institute of Technology (AIT)",
+  "Acharya Narendra Dev College, University Of Delhi",
+  "Achievement House Charter School - Online",
+  "Acropolis Institute of Technology & Research",
+  "Acton-Boxborough Regional High School",
+  "Adelphi University",
+  "Aditya Institute of Technology and Management (AITAM College, Tekkali)",
+  "Adlai E. Stevenson High School",
+  "Advanced Math and Science Academy Charter School",
+  "Agnes Scott College",
+  "Agora Cyber Charter School",
+  "Alagappa Chettiar Government College of Engineering and Technology",
+  "Alagappa College of Technology, Anna University",
+  "Alameda High School",
+  "Albany Medical College",
+  "Albany State University (GA)",
+  "Albertian Institute of Science and Technology (AISAT)",
+  "Albright College",
+  "Alfa College",
+  "Aligarh Muslim University",
+  "Allen High School",
+  "Alwar Institute of Engineering and Technology (AIET)",
+  "Ambala College of Engineering and Applied Research",
+  "Ambedkar Institute of Advanced Communication Technologies and Research (AIACTR)",
+  "American Heritage School",
+  "American High School",
+  "American River College, California",
+  "American University in Dubai",
+  "American University, Washington, D.C.",
+  "Amherst College",
+  "Amity School of Engineering and Technology",
+  "Amity University",
+  "Amrita School of Engineering",
+  "Amritsar College of Engineering & Technology",
+  "Anand Institute of Higher Technology",
+  "Ancaster High School",
+  "Anchor Bay High School",
+  "Andhra University College of Engineering",
+  "Andover Central High School",
+  "Angadi Institute of Technology & Management (AITM)",
+  "Anil Neerukonda Institute of Technology and Sciences",
+  "Anjalai Ammal Mahalingam Engineering College",
+  "Anna University",
+  "Ansal Technical Campus, Dr. A.P.J Abdul Kalam Technical University",
+  "Anurag University, Ghatkesar",
+  "Apeejay Stya University",
+  "Appalachian State University",
+  "Aravali Institute of Technical Studies",
+  "Arcadia High School, California",
+  "Arcadia University",
+  "Arizona State University",
+  "Army Institute Of Technology, Pune",
+  "Art Institute of Philadelphia",
+  "Arya College of Engineering & I.T.",
+  "Asansol Engineering College",
+  "Ashoka Institute of Technology and Management",
+  "Asia Pacific Institute of Information Technology, Panipat",
+  "Asia Pacific University of Information & Technology, Kuala Lumpur",
+  "Asian School of Business Management (ASBM University)",
+  "Assam Downtown University",
+  "Assam Engineering College",
+  "Assam University, Silchar",
+  "Aston University",
+  "Atal Bihari Vajpayee Indian Institute of Information Technology and Management, Gwalior (ABV-IIITM Gwalior)",
+  "Atlanta Metropolitan State College",
+  "Atlantic Cape Community College",
+  "Atma Ram Sanatan Dharma College",
+  "Atria Institute of Technology",
+  "Auburn University",
+  "Audisankara College of Engineering and Technology",
+  "Aurora Group of Institutions",
+  "Austin Community College District",
+  "Aviation Career & Technical Education High School",
+  "Avon High School",
+  "B. P. Poddar Institute of Management and Technology",
+  "B. V. Bhoomaraddi College of Engineering and Technology (KLE Tech)",
+  "B.M.S College Of Engineering",
+  "B.N.M Institute of Technology",
+  "BITS Pilani, Hyderabad Campus",
+  "BITS Pilani, K K Birla Goa Campus",
+  "BLDEA’s V.P. Dr P. G. Halakatti College of Engineering & Technology",
+  "BMIET, Sonipat",
+  "BMIIT, Uka Tarsadia University, Bardoli, Surat",
+  "BML Munjal University (BMU)",
+  "BMS Institute of Technology and Management",
+  "BVRIT Hyderabad College of Engineering for Women",
+  "Babaria Institute of Technology",
+  "Babson College",
+  "Babu Banarasi Das National Institute of Technology and Management",
+  "Babu Banarasi Das Northern India Institute of Technology",
+  "Babu Banarsi Das Institute of Technology",
+  "Badruka Educational Society",
+  "Bahria University Lahore Campus",
+  "Ball State University",
+  "Baltimore Polytechnic Institute",
+  "Bangalore Institute of Technology",
+  "Bangalore University",
+  "Bannari Amman Institute of Technology",
+  "Bapuji Institute Of Engineering & Technology (BIET)",
+  "Bard College",
+  "Barnard College",
+  "Barton College",
+  "Baruch College, CUNY",
+  "Basaveshwar Engineering College",
+  "Baton Rouge Community College",
+  "Battlefield High School",
+  "Bauman Moscow State Technical University",
+  "Bayside High School",
+  "Bayview Secondary School",
+  "Beihang University",
+  "Bellevue College, Washington",
+  "Benedictine College",
+  "Benha University",
+  "Benjamin Franklin High School - Baltimore",
+  "Benjamin Franklin High School - Philadelphia",
+  "Bennett College",
+  "Bennett University (Times of India Group)",
+  "Bentley University",
+  "Berea College",
+  "Bergen Catholic High School",
+  "Bergen Community College",
+  "Bergen County Academies",
+  "Bergen County Technical High School - Teterboro",
+  "Berkshire Community College",
+  "Bhagalpur College of Engineering",
+  "Bhagwan Parshuram Institute of Technology",
+  "Bharat Institute of Engineering and Technology (BIET)",
+  "Bharathiar University",
+  "Bharati Vidyapeeth's College of Engineering",
+  "Bilkent University",
+  "Bineswar Brahma Engineering College (BBEC)",
+  "Binghamton University",
+  "Birkbeck, University of London",
+  "Birla Institute Of Technology,  Mesra",
+  "Birla Institute of Technology and Science, Pilani",
+  "Birla Institute of Technology, Patna",
+  "Birla Vishvakarma Mahavidyalaya Engineering College",
+  "Birmingham City University",
+  "Birsa Institute of Technology (BIT), SINDRI",
+  "Blinn College",
+  "Bloomfield Hills High School",
+  "Bloomsburg University of Pennsylvania",
+  "Blue Mountain Academy",
+  "BlueCrest University College",
+  "Bluevale Collegiate Institute",
+  "Boca Raton Community High School",
+  "Boise State University",
+  "Bordentown Regional High School",
+  "Borough of Manhattan Community College, CUNY",
+  "Boston College",
+  "Boston Latin School",
+  "Boston University",
+  "Boston University Metropolitan College",
+  "Bourne Grammar School",
+  "Bournemouth University",
+  "Bowdoin College",
+  "Bowie State University",
+  "Boys Latin of Philadelphia Charter School",
+  "Brampton Centennial Secondary School",
+  "Brandeis University",
+  "Brentsville High School",
+  "Briar Cliff University",
+  "Briarcliff High School",
+  "Bridgewater State University",
+  "Brigham Young University",
+  "British Columbia Institute of Technology",
+  "Brno University of Technology",
+  "Brock University",
+  "Bronx Community College, CUNY",
+  "Brookdale Community College",
+  "Brooklyn College, CUNY",
+  "Brooklyn Technical High School",
+  "Brookwood High School",
+  "Brown University",
+  "Brunel University London",
+  "Bryn Athyn College",
+  "Bryn Mawr College",
+  "Bucknell University",
+  "Bucks County Community College",
+  "Bundelkhand Institute Of Engineering & Technology (BIET Jhansi)",
+  "Burlington Township High School",
+  "Business Academy Aarhus",
+  "C. Abdul Hakeem College of Engineering & Technology",
+  "C. D. Hylton High School",
+  "C. K. Pithawala College of Engineering and Technology",
+  "C.V. Raman College of Engineering",
+  "CAPA - Philadelphia High School for Creative and Performing Arts",
+  "CMR College of Engineering and Technology,  Hyderabad",
+  "CMR Engineering College",
+  "CMR Institute of Technology (CMRIT)",
+  "CMR Technical Campus",
+  "CODE University of Applied Sciences Berlin",
+  "COMSATS Institute of Information Technology",
+  "Cabrini University",
+  "Cadbury Sixth Form College",
+  "Cairn University",
+  "Caldwell University",
+  "California High School",
+  "California Institute of Technology",
+  "California Polytechnic State University, San Luis Obispo",
+  "California State Polytechnic University, Pomona",
+  "California State University, Bakersfield",
+  "California State University, Channel Islands",
+  "California State University, Chico",
+  "California State University, Dominguez Hills",
+  "California State University, East Bay",
+  "California State University, Fresno",
+  "California State University, Fullerton",
+  "California State University, Humboldt",
+  "California State University, Long Beach",
+  "California State University, Los Angeles",
+  "California State University, Maritime",
+  "California State University, Monterey Bay",
+  "California State University, Northridge",
+  "California State University, Sacramento",
+  "California State University, San Bernardino",
+  "California State University, San Diego",
+  "California State University, San Francisco",
+  "California State University, San Jose",
+  "California State University, San Luis Obispo",
+  "California State University, San Marcos",
+  "California State University, Sonoma",
+  "California State University, Stanislaus",
+  "California University of Pennsylvania",
+  "Calvin College",
+  "Camden County College",
+  "Cameron Heights Collegiate Institute",
+  "Canada (Cañada) College",
+  "Canara Engineering College (CEC)",
+  "Canyon Crest Academy",
+  "Cardiff Metropolitan University",
+  "Carleton College",
+  "Carleton University",
+  "Carnegie Mellon University",
+  "Carteret High School",
+  "Carthage College",
+  "Cascadia College",
+  "Case Western Reserve University",
+  "Cathedral High School, Los Angeles",
+  "Catholic University of America",
+  "Cedar Creek High School",
+  "Cedar Ridge High School",
+  "Cedarville University",
+  "Centennial Collegiate Vocational Institute",
+  "Centennial High School",
+  "Central Connecticut State University",
+  "Central High School - Philadelphia",
+  "Central Institute of Plastics Engineering & Technology (CIPET)",
+  "Central PA Digital Learning Foundation Charter School",
+  "Central Peel Secondary School",
+  "Central Texas College",
+  "Centro de Enseñanza Técnica y Superior (CETYS), Campus Ensenada",
+  "Centro de Enseñanza Técnica y Superior (CETYS), Campus Mexicali",
+  "Cerritos College",
+  "Chaitanya Bharathi Institute of Technology",
+  "Chalmers University of Technology",
+  "Champlain College",
+  "Chandigarh College Of Engineering & Technology (CCET)",
+  "Chandigarh University",
+  "Channabasaveshwara Institute of Technology",
+  "Chaparral Star Academy",
+  "Chapel Hill High School",
+  "Charotar University Of Science And Technology (CHAURSAT)",
+  "Charter High School for Architecture and Design - Philadelphia",
+  "Chattahoochee Technical College",
+  "Cherokee High School",
+  "Cherry Hill High School East",
+  "Cherry Hill High School West",
+  "Chestnut Hill College",
+  "Cheyney University",
+  "Chinguacousy Secondary School",
+  "Chitkara Institute of Engineering & Technology (CIET)",
+  "Chitkara University",
+  "Christ College of Engineering and Technology",
+  "Christ Knowledge City",
+  "Christ University",
+  "Christ University Faculty of Engineering",
+  "Cincinnati State Technical and Community College",
+  "Citrus College",
+  "City College of San Francisco",
+  "City Engineering College",
+  "City Neighbors High School",
+  "City University London",
+  "Claremont McKenna College",
+  "Clarion University of Pennsylvania",
+  "Clark Atlanta University",
+  "Clark University",
+  "Clarksburg High School",
+  "Clarkson University",
+  "Clayton State University",
+  "Clemson University",
+  "Cleveland State University",
+  "Clifton Public Highschool",
+  "Cluster Innovation Centre, University of Delhi",
+  "Cochin College of Engineering and Technology",
+  "Cochin University College of Engineering Kuttanad",
+  "Cochin University of Science And Technology",
+  "Cochin University of Science and Technology",
+  "Coe College",
+  "Coimbatore Institute of Engineering and Technology (CIET)",
+  "Coimbatore Institute of Technology (CIT)",
+  "Colegio Simón Bolívar",
+  "Colgate University",
+  "College of Agricultural Engineering and Post Harvest Technology (CAEPHT)",
+  "College of Agriculture, Central Agricultural University",
+  "College of Charleston",
+  "College of DuPage",
+  "College of Engineering & Management Punnapra",
+  "College of Engineering Chengannur",
+  "College of Engineering and Management, Kolaghat",
+  "College of Engineering, Pune",
+  "College of Staten Island, CUNY",
+  "College of Technology & Engineering, Udaipur",
+  "College of Westchester",
+  "Colleyville Heritage High School",
+  "Collins Hill High School",
+  "Collège Ahuntsic",
+  "Collège André-Grasset",
+  "Collège Français",
+  "Collège Jean-Eudes",
+  "Collège Jean-de-Brébeuf",
+  "Collège Lionel-Groulx",
+  "Collège Regina Assumpta",
+  "Collège de Bois-de-Boulogne",
+  "Collège de Maisonneuve",
+  "Collège de Montréal",
+  "Collège de Rosemont",
+  "Colorado School of Mines",
+  "Colts Neck High School",
+  "Columbia Secondary School",
+  "Columbia University",
+  "Columbus College of Art and Design",
+  "Columbus State Community College",
+  "Comenius University",
+  "Commonwealth Charter Academy Charter School",
+  "Community Academy of Philadelphia Charter School",
+  "Community College of Allegheny County",
+  "Community College of Baltimore County",
+  "Community College of Philadelphia",
+  "Community College of Rhode Island",
+  "Concord Academy",
+  "Concordia University",
+  "Concordia University Ann Arbor",
+  "Concordia University Chicago",
+  "Concordia University Irvine",
+  "Concordia University Nebraska",
+  "Concordia University St. Paul",
+  "Concordia University Texas",
+  "Concordia University Wisconsin",
+  "Conestoga College",
+  "Conestoga High School",
+  "Connecticut College",
+  "Conroe ISD Academy of Science and Technology, Texas",
+  "Constitution High School - Philadelphia",
+  "Cooch Behar Government Engineering College",
+  "Cooper Union",
+  "Coral Glades High School",
+  "Cornell College",
+  "Cornell University",
+  "Council Rock High School North",
+  "Council Rock High School South",
+  "County College of Morris",
+  "Covenant University",
+  "Coventry University",
+  "Cranbrook Schools",
+  "Cranfield University",
+  "Creekview High School",
+  "Cumberland County College",
+  "Cummins College of Engineering for Women,  Pune",
+  "Cupertino High School",
+  "Cégep André-Laurendeau",
+  "Cégep Marie-Victorin",
+  "Cégep de Saint-Laurent",
+  "Cégep du Vieux Montréal",
+  "D.J. College of Engineering & Technology",
+  "D.K.T.E Society's Textile and Engineering Institute",
+  "DCS,  Ganpat University",
+  "DIT University",
+  "Dalhousie University",
+  "Dalmia Institute of Scientific & Industrial Research",
+  "Dartmouth College",
+  "Davidson College",
+  "Dawson College",
+  "Dayalbagh Educational Institute",
+  "Dayananda Sagar University",
+  "De Anza College",
+  "DePaul University",
+  "DePauw University",
+  "DeSales University",
+  "Deenbandhu Chhotu Ram University of Science and Technology, Murthal",
+  "Deerfield High School",
+  "Del Norte High School",
+  "Delaware County Community College - Downingtown",
+  "Delaware County Community College - Exton",
+  "Delaware County Community College - Main Campus (Marple)",
+  "Delaware County Community College - Phoenixville",
+  "Delaware County Community College - Sharon Hill",
+  "Delaware County Community College - Upper Darby",
+  "Delaware County Community College - West Grove",
+  "Delaware State University",
+  "Delaware Technical Community College",
+  "Delaware Valley Academy of Medical and Dental Assistants",
+  "Delaware Valley University",
+  "Delft University of Technology",
+  "Delhi Technological University",
+  "Denison University",
+  "Department of Human Resource Management and OB, Central University of Jammu",
+  "Des Moines Area Community College",
+  "Devry University - Philadelphia Center City",
+  "Dharmsinh Desai University",
+  "Dhirubhai Ambani Institute of Information and Communication Technology (DA-IICT)",
+  "Diablo Valley College",
+  "Dickinson College",
+  "Digital Harbor High School",
+  "Don Bosco College of Engineering",
+  "Don Bosco College of Engineering and Technology",
+  "Don Bosco Institute of Technology",
+  "Doon College of Engineering & Technology",
+  "Dougherty Valley High School",
+  "Dr B. R. Ambedkar Institute of Technology,  Port Blair",
+  "Dr. A.P.J. Abdul Kalam Technical University, Lucknow",
+  "Dr. Akhilesh Das Gupta Institute of Technology & Management",
+  "Dr. B. R. Ambedkar National Institute of Technology Jalandhar",
+  "Dr. B.C. Roy Engineering College, Durgapur",
+  "Dr. Babasaheb Ambedkar Marathwada University",
+  "Dr. Harisingh Gour University, Sagar University",
+  "Dr. K.N. Modi Engineering College",
+  "Dr. MGR Educational Research Institute University",
+  "Dr. SJS Paul Memorial College of Engineering and Technology (CIT)",
+  "Dr. T. Thimmaiah Institute of Technology",
+  "Drake University",
+  "Drew University",
+  "Drexel University",
+  "Dublin High School",
+  "Dublin Jerome High School",
+  "Duke University",
+  "Dulaney High School",
+  "Duquesne University",
+  "Durant High School",
+  "Durham College",
+  "Durham University",
+  "Dwarkadas J. Sanghvi College of Engineering",
+  "Dwight-Englewood School",
+  "EASTERN Center for Arts and Technology",
+  "EPFL | École polytechnique fédérale de Lausanne",
+  "EPITECH Bordeaux",
+  "Earl of March Secondary School",
+  "Earlham College",
+  "East Brunswick High School",
+  "East Central University",
+  "East Chapel Hill High Schoo",
+  "East Los Angeles College",
+  "East Point College of Engineering and Technology",
+  "East Stroudsburg High School",
+  "East West Institute of Technology",
+  "Eastern High School - Louisville",
+  "Eastern Michigan University",
+  "Eastern Regional High School",
+  "Eastern University - St. Davids",
+  "Eastern University Academy Charter School",
+  "Eastern Washington University",
+  "Eckerd College",
+  "Edina High School",
+  "Edinburgh Napier University",
+  "Edison Academy",
+  "Edison High School",
+  "Edward R. Murrow High School",
+  "Egg Harbor Township High School",
+  "Eidgenössische Technische Hochschule (ETH) Zürich",
+  "Ekta Incubation Center, West Bengal",
+  "El Camino College",
+  "El Centro College",
+  "El Centro de Estudiantes",
+  "Elgin Academy",
+  "Elizabeth High School",
+  "Elon University",
+  "Embry-Riddle Aeronautical University",
+  "Emory University",
+  "Entrepreneurship Development Center, MIT, Pune",
+  "Entreprpeneurship Development Cell, University of Kerala",
+  "Episcopal Academy",
+  "Er.Perumal Manimekalai College of Engineering",
+  "Erasmus Hogeschool Brussel",
+  "Erie Community College",
+  "Ernest Manning High School",
+  "Esperanza Academy Charter School",
+  "Esperanza Cyber Charter School",
+  "Evergreen Valley College",
+  "Evergreen Valley High School",
+  "Fachhochschule Dortmund",
+  "Faculty Of Engineering & Technology, Gurukula Kangri Vishwavidyalaya",
+  "Faculty of science / Ibn Tofail University",
+  "Fahaheel Al-Watanieh Indian Private School",
+  "Fairfield University",
+  "Fairleigh Dickinson University",
+  "Fairview High School",
+  "Farmingdale State College",
+  "FernUniversität in Hagen",
+  "Finolex Academy of Management and Technology",
+  "First Philadelphia Preparatory Charter School",
+  "Fitchburg State University",
+
+  "Fontys Hogeschool",
+  "Foothill College",
+  "Fordham University",
+  "Forest Heights Collegiate Institute",
+  "Forest Park High School - Baltimore",
+  "Forest Park High School - Forest Park, GA",
+  "Forest Park High School - Woodbridge",
+  "Fort Scott Community College",
+  "Foundation Collegiate Academy",
+  "Foundation for Innovation and Technology Transfer, IIT Delhi",
+  "Fr. Conceicao Rodrigues College of Engineering",
+  "Francis Holland School",
+  "Francis Lewis High School",
+  "Frankford High School - Philadelphia",
+  "Franklin High School",
+  "Franklin Learning Center - Philadelphia",
+  "Franklin Towne Charter High School",
+  "Franklin W. Olin College of Engineering",
+  "Frederick Community College",
+  "Freedom High School - Bethlehem",
+  "Freedom High School - Woodbridge",
+  "Freehold High School",
+  "Freire Charter High School",
+  "Fremont High School",
+  "Full Sail University",
+  "Fullerton College",
+  "G. H. Patel College of Engineering & Technology",
+  "G. Narayanamma Institute of Technology Science (For Women)",
+  "G.H. Raisoni College of Engineering",
+  "GEC, Gandhinagar",
+  "GEC, Patan",
+  "GIDC Degree Engineering College",
+  "GITAM Centre for Integrated Rural Development",
+  "GL Bajaj Institute of Technology and Management",
+  "GMR Institute of Technology",
+  "GOA IT INNOVATION CENTRE",
+  "GSSS Institute of Engineering & Technology for Women",
+  "GZS Campus College of Engineering & Technology",
+  "Galgotias College of Engineering & Technology",
+  "Gandhi Institute of Technical Advancement (GITA)",
+  "Gandhi Institute of Technology and Management, Bengaluru",
+  "Gandhi Institute of Technology and Management, Hyderabad",
+  "Gandhi Institute of Technology and Management, Visakhapatnam",
+  "Gandhi Institution of Management Studies",
+  "Ganga International School",
+  "Ganpat University",
+  "Gar-Field Senior High School",
+  "Garnet Valley High School",
+  "Gautam Buddha University",
+  "Gaya College Of Engineering",
+  "Gayatri Vidya Parishad College of Engineering",
+  "Geetanjali Institute of Technical Studies (GITS)",
+  "Geethanjali College of Engineering and Technology",
+  "George C. Marshall High School",
+  "George Heriot's School",
+  "George Mason University",
+  "George Washington High School - Philadelphia",
+  "Georgetown University",
+  "Georgia Institute of Technology",
+  "Georgia State University",
+  "Germantown Friends School",
+  "Geroge Washington Carver High School - Philadelphia",
+  "Ghent University",
+  "Ghousia College of Engineering",
+  "Girijananda Chowdhury Institute of Management and Technology (GIMT)",
+  "Gitam School of Technology",
+  "Glassboro High School",
+  "Glenaeon Rudolf Steiner School",
+  "Glenbrook North High School",
+  "Glenbrook South High School",
+  "Glendale Community College",
+  "Glenforest Secondary School",
+  "Global Academy of Technology",
+  "Goa College of Engineering",
+  "Gokaraju Rangaraju Institute of Engineering and Technology (GRIET)",
+  "Goldsmiths, University of London",
+  "Gopalan College of Engineering and Management",
+  "Gordon Graydon Memorial Secondary School",
+  "Gottfried Wilhelm Leibniz Universität Hannover",
+  "Government College Of Engineering, Amravati",
+  "Government College Of Engineering, Aurangabad",
+  "Government College Of Engineering, Karad",
+  "Government College of Engineering & Technology, Jammu",
+  "Government College of Engineering, Bargur",
+  "Government College of Engineering, Kalahandi",
+  "Government College of Engineering, Kannur",
+  "Government College of Engineering, Salem",
+  "Government College of Technology, Coimbatore",
+  "Government Engineering College Palakkad, Sreekrishnapuram",
+  "Government Engineering College, Ajmer",
+  "Government Engineering College, Banswara",
+  "Government Engineering College, Hassan",
+  "Government Engineering College, Kozhikode",
+  "Government Engineering College, Thrissur",
+  "Government Model Engineering College, Thrikkakara",
+  "Government Polytechnic Gandhinagar",
+  "Government Sri Krishnarajendra Silver Jubilee Technological Institute",
+  "Governor's School for Science & Technology",
+  "Govind Ballabh Pant Institute of Engineering & Technology",
+  "Grady High School",
+  "Grand Rapids Community College",
+  "Grand Valley State University",
+  "Graphic Era University",
+  "Great Neck South High School",
+  "Greater Lowell Technical High School",
+  "Green River College",
+  "Greenwood College School",
+  "Grinnell College",
+  "Guelph Collegiate Vocational Institute",
+  "Gujarat Energy Research and Management Institute (GERMI)",
+  "Gujarat Technological University",
+  "Gujarat University",
+  "Guru Ghasidas Vishwavidyalaya, Bilaspur",
+  "Guru Gobind Singh Indraprastha University",
+  "Guru Jambheshwar University of Science and Technology (GJUS&T), HISAR",
+  "Guru Jambheshwar University of Science and Technology, Hisar",
+  "Guru Nanak Dev Engineering College",
+  "Guru Nanak Institutions",
+  "Guru Tegh Bahadur Institute of Technology (GTBIT)",
+  "Gurukula Kangri University",
+  "Guttman Community College, CUNY",
+  "Gwalior Engineering College",
+  "Gwinnett Technical College",
+  "Gwynedd Mercy University",
+  "H.N. Werkman College",
+  "HAN University of Applied Sciences",
+  "HKBK College of Engineering",
+  "HMR Institute of Technology & Management, GGSIPU",
+  "HMS Institute of Technology",
+  "Haaga-Helia University of Applied Sciences",
+  "Haldia Institute of Technology",
+  "Hamilton College",
+  "Hamline University",
+  "Hampshire College",
+  "Hampton University",
+  "Hanze University of Applied Sciences",
+  "Harcourt Butler Technical University, Kanpur",
+  "Harcum College",
+  "Harper College",
+  "Harrisburg Area Community College",
+  "Harrisburg University - Harrisburg Campus",
+  "Harrisburg University - Philadelphia Campus",
+  "Harrison Career Institute",
+  "Harvard Medical School",
+  "Harvard University",
+  "Harvey Mudd University",
+  "Haryana Engineering College",
+  "Hasso-Plattner-Institut Academy",
+  "Haverford College",
+  "Hazleton Area High School",
+  "Head-Royce School",
+  "Health Careers High School",
+  "Heartland Community College",
+  "Helwan University",
+  "Henry M. Gunn High School",
+  "Herguan University",
+  "Heritage Institute of Technology",
+  "Het Baarnsch Lyceum",
+  "Hi-Tech Institute of Engineering & Technology",
+  "Hi-Tech Institute of Technology",
+  "High Technology High School",
+  "Highland Park High School",
+  "Hightstown High School",
+  "Hillsborough Community College",
+  "Hillsborough High School",
+  "Hindustan College of Science & Technology",
+  "Hindustan Institute of Technology & Science",
+  "Hinsdale Central High School",
+  "Hiram College",
+  "Hirasugar Institute of Technology, Nidasoshi",
+  "Hofstra University",
+  "Hogeschool Thomas More",
+  "Hogeschool van Amsterdam",
+  "Holton-Arms School",
+  "Holy Family University",
+  "Homestead High School",
+  "Hong Kong University of Science and Technology",
+  "Hood College",
+  "Horace Furness High School",
+  "Horace Mann School",
+  "Hostos Community College, CUNY",
+  "Houghton High School",
+  "Houston Community College",
+  "Howard University",
+  "Hudson County Community College",
+  "Hudson Valley Community College",
+  "Hunter College High School",
+  "Hunter College, CUNY",
+  "Huron Heights Secondary School",
+  "Hussian School of Art",
+  "I.K. Gujral Punjab Technical University Jalandhar (IKGPTU)",
+  "I.T.S Engineering College",
+  "IAN Mentoring and Incubation Services",
+  "IIMT College Of Medical Sciences, Meerut",
+  "IIMT College of Engineering, Greater Noida",
+  "IIMT College of Pharmacy, Greater Noida",
+  "IIMT Engineering College, Meerut",
+  "IKP Knowledge Park Erstwhile ICICI Knowledge Park",
+  "IMS Engineering College",
+  "IT University of Copenhagen",
+  "ITM University, Gwalior",
+  "ITM University, Vadodara",
+  "ITMO University",
+  "Iliria College",
+  "Illinois Institute of Technology",
+  "Illinois State University",
+  "Imhotep Institute Charter High School",
+  "Immaculata University",
+  "Impact College of Engineering and Applied Science",
+  "Imperial College London",
+  "Inderprastha Engineering College (IPEC)",
+  "Indian Hills Community College",
+  "Indian Institute of Engineering Science and Technology (IIEST), Shibpur",
+  "Indian Institute of Information Technology Design & Manufacturing, Jabalpur",
+  "Indian Institute of Information Technology, Allahabad",
+  "Indian Institute of Information Technology, Kalyani",
+  "Indian Institute of Information Technology, Kottayam",
+  "Indian Institute of Information Technology, Pune",
+  "Indian Institute of Information Technology, Sri City",
+  "Indian Institute of Information Technology, Una",
+  "Indian Institute of Information Technology, Vadodara",
+  "Indian Institute of Space Science and Technology (IIST)",
+  "Indian Institute of Technology (ISM), Dhanbad",
+  "Indian Institute of Technology, BHU",
+  "Indian Institute of Technology, Bhubaneswar",
+  "Indian Institute of Technology, Bombay",
+  "Indian Institute of Technology, Gandhinagar",
+  "Indian Institute of Technology, Guwahati",
+  "Indian Institute of Technology, Gwalior",
+  "Indian Institute of Technology, Hyderabad",
+  "Indian Institute of Technology, Jabalpur",
+  "Indian Institute of Technology, Jodhpur",
+  "Indian Institute of Technology, Kanpur",
+  "Indian Institute of Technology, Kharagpur",
+  "Indian Institute of Technology, Kota",
+  "Indian Institute of Technology, Madras",
+  "Indian Institute of Technology, Patna",
+  "Indian Institute of Technology, Roorkee",
+  "Indian Institute of Technology, Ropar",
+  "Indiana State University",
+  "Indiana University",
+  "Indiana University of Pennsylvania",
+  "Indiana University-Purdue University Fort Wayne",
+  "Indiana University–Purdue University Indianapolis",
+  "Indira Gandhi Delhi Technical University for Women",
+  "Indira Gandhi Engineering College, Sagar",
+  "Indira Gandhi Institute of Technology, Sarang",
+  "Indira Gandhi National Open University",
+  "Indraprastha Institute of Information Technology",
+  "Indus University, Ahmedabad",
+  "Insight PA Cyber Charter School",
+  "Institut polytechnique de Bordeaux (INP)",
+  "Institute Of Technology, Nirma University",
+  "Institute for Auto Parts and Hand Tools Technology",
+  "Institute of Aeronautical Engineering (IARE), Hyderabad",
+  "Institute of Engineering & Management (IEM)",
+  "Institute of Engineering and Rural Technology Allahabad",
+  "Institute of Engineering and Technology, DAVV",
+  "Institute of Infrastructure Technology Research and Management, Ahmedabad",
+  "Institute of Technical Education and Research (ITER), Bhubaneswar",
+  "Institute of Technology, Banaras Hindu University",
+  "Instituto Politécnico Nacional",
+  "Instituto Tecnologico Superior de San Martin Texmelucan",
+  "Instituto Tecnológico Autónomo de México (ITAM)",
+  "Instituto Tecnológico Superior de Cintalapa",
+  "Instituto Tecnológico Superior de El Mante",
+  "Instituto Tecnológico Superior de los Ríos",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM)",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Aguascalientes",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Chiapas",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Chihuahua",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Ciudad Juárez",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Ciudad de Mexico",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Cuernavaca",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Cumbres",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Eugenio Garza Lagüera",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Eugenio Garza Sada",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Guadalajara",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Hidalgo",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Irapuato",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Laguna",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus León",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Morelia",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Obregón",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Puebla",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Querétaro",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Saltillo",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus San Luis Potosí",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Santa Catarina",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Santa Fe",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Sinaloa",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Sonora",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Tampico",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Toluca",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Valle Alto",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Veracruz",
+  "Instituto Tecnológico y de Estudios Superiores de Monterrey (ITESM) Campus Zacatecas",
+  "Instituto Tecnológico y de Estudios Superiores de Occidente (ITESO)",
+  "Instituto Tecnólogico de La Laguna (ITL)",
+  "Instytut Pamięci Narodowej",
+  "International Institute Of Information Technology, Naya Raipur",
+  "International Institute of Information Technology,  Hyderabad",
+  "International Institute of Information Technology, Bangalore",
+  "International Leadership Charter High School",
+  "International School of Choueifat",
+  "Iowa Central Community College",
+  "Iowa State University",
+  "Iowa Western Community College",
+  "Islamic University of Science and Technology, Pulwama",
+  "Istanbul University",
+  "Ithaca College",
+  "J.C. Bose University of Science and Technology, YMCA",
+  "J.N.N College of Engineering",
+  "JK Institute of Applied Physics and Technology",
+  "JK Lakshmipat University (JKLU)",
+  "JNTUA College of Engineering,  Pulivendula",
+  "JNTUH College of Engineering, HYDERABAD",
+  "JNTUK University College of Engineering, Vizianagaram",
+  "JSS Academy of Technical Education",
+  "Jabalpur Engineering College",
+  "Jackson Memorial High School",
+  "Jackson State University",
+  "Jacobs University Bremen",
+  "Jadavpur University",
+  "Jagiellonian University",
+  "Jai Narain Vyas University",
+  "Jaipur Engineering College & Research Centre (JECRC)",
+  "Jaipur National University",
+  "Jalpaiguri Government Engineering College",
+  "James Gillespie's High School",
+  "James Madison High School",
+  "James Madison University",
+  "Jamia Hamdard",
+  "Jamia Millia Islamia - JMI,  Jamia Nagar",
+  "Jawaharlal Nehru Government Engineering College (JNGEC)",
+  "Jawaharlal Nehru Technological University",
+  "Jaypee Institute of Technology,  Noida",
+  "Jaypee University of Engineering and Technology",
+  "Jerusalem College of Engineering",
+  "Jnanavikas Institute of Technology",
+  "Jodhpur Institute of Engineering and Technology (JIET)",
+  "John A. Ferguson Senior High School",
+  "John Abbott College",
+  "John Bartram High School",
+  "John F. Kennedy Memorial High School",
+  "John Jay College of Criminal Justice, CUNY",
+  "John Leggott College",
+  "John P. Stevens High School",
+  "Johns Hopkins University",
+  "Johnson & Wales University",
+  "Johnson C. Smith University",
+  "Jorhat Engineering College",
+  "Jorhat Institute of Science and Technology",
+  "Jules E. Mastbaum Technical High School",
+  "Julia R. Masterman School",
+  "Jyothy Institute of Technology",
+  "K S School of Engineering and Management",
+  "K. J. Somaiya College of Engineering",
+  "K. S Institute of Technology (KSIT), Bengaluru",
+  "K.L. College of Engineering",
+  "K.L.S Gogte Institute of Technology",
+  "K.M.E.A Engineering College",
+  "K.S Rangasamy College Of Technology",
+  "K.S. School of Business Management",
+  "KCG College of Engineering",
+  "KIPP DuBois Charter School",
+  "KJ's Educational Institutes,  Pune",
+  "KLE Dr. M.S. Sheshgiri College of Engineering and Technology",
+  "KLN College of Engineering",
+  "KLS Gogte Institute of Technology",
+  "KNSIT",
+  "KTH Royal Institute of Technology",
+  "Kamla Nehru Institute of Technology",
+  "Kansai University",
+  "Kansas State University",
+  "Kantipur Engineering College",
+  "Karlsruhe Institute of Technology",
+  "Karmaveer Bhaurao Patil College of Engineering",
+  "Karpagam College of Engineering (KCE)",
+  "Karunya Institute of Technology and Sciences",
+  "Kashi Institute of Technology",
+  "Kathmandu BernHardt College",
+  "Kaunas University of Technology",
+  "Kean University",
+  "Keele University",
+  "Kendriya Vidyalaya, AFS, Begumpet",
+  "Kennesaw State University",
+  "Kennett High School",
+  "Kensington High School Complex",
+  "Kent State University",
+  "Kent State University at Stark",
+  "Keshav Memorial Institute of Technology,  Hyderabad",
+  "Khan Lab School",
+  "King Edward VI Five Ways School",
+  "King's College London",
+  "Kingsborough Community College, CUNY",
+  "Kingsway Regional High School",
+  "Kitchener-Waterloo Collegiate & Vocational School",
+  "Knox College",
+  "Konark Institute of Science and Technology",
+  "Kongu Engineering College",
+  "Koustuv Group Of Institutions (KISD & COEB)",
+  "Kraków University of Economics",
+  "Krishi Vigyan Kendra, Durgapur",
+  "Krishna Engineering College",
+  "Kristu Jayanti College",
+  "Kshatriya College of Engineering",
+  "Kumaraguru College Of Technology",
+  "Kutztown University of Pennsylvania",
+  "L D College Of Engineering Library",
+  "L. D. College of Engineering",
+  "LBS Institute of Technology for Women (LBSITW)",
+  "LICET",
+  "LIM College",
+  "La Roche College",
+  "La Salle University - Philadelphia",
+  "La Sierra University",
+  "LaGuardia Community College, CUNY",
+  "Lady Doak College",
+  "Lafayette College",
+  "Lake Braddock Secondary School",
+  "Lakeside High School",
+  "Lakshmi Narayan College of Technology (LNCT)",
+  "Lampeter-Strasburg High School",
+  "Lancaster University",
+  "Lankenau High School",
+  "Laval University",
+  "Lawrence Technological University",
+  "Lawrence University",
+  "Lehigh University",
+  "Lehman College, CUNY",
+  "Leiden University",
+  "Lewis & Clark College",
+  "Lewis University",
+  "Lexington High School",
+  "Lick Wilmerding High School",
+  "Lincoln Christian University",
+  "Lincoln Technical Institute - Center City Philadelphia",
+  "Lincoln Technical Institute - Northeast Philadelphia",
+  "Lincoln University",
+  "Lindenwood University",
+  "Linn-Mar High School",
+  "Lisgar Collegiate Institute",
+  "Little Flowers Public Sr Secondary School",
+  "Livingston High School",
+  "Loch Raven High School",
+  "Lodz University of Technology",
+  "Loknayak Jai Prakash Institute of Technology,  Chhapra",
+  "London Metropolitan University",
+  "London School of Economics and Political Science",
+  "Lone Star College System",
+  "Lord Krishna College of Engineering",
+  "Lords Institute of Engineering & Technology",
+  "Los Altos High School",
+  "Loughborough University",
+  "Louisiana State University",
+  "Lovely Professional University",
+  "Lowell High School",
+  "Loyola Marymount University",
+  "Luleå University of Technology, LTU",
+  "Luther College",
+  "Lyallpur Khalsa College of Engineering, Jalandhar",
+  "Lynbrook High School",
+  "M.J.P. Rohilkhand University",
+  "M.S. Ramaiah School of Advance Studies",
+  "M.V.Jayaraman College of Engineering",
+  "MBM Engineering College, Jodhpur",
+  "MSME TDC PPDC Agra",
+  "MVJ College of Engineering",
+  "MacArthur High School",
+  "MacEwan University",
+  "Macalester College",
+  "Macaulay Honors College, CUNY",
+  "Macomb Community College",
+  "Madan Mohan Malaviya University of Technology, Gorakhpur",
+  "Madhav Institute of Technology & Science (MITS)",
+  "Madison College",
+  "Madison West High School",
+  "Madras Institute Of Technology",
+  "Maggie L. Walker Governor's School",
+  "Mahakal Institute Of Technology",
+  "Maharaj Vijayaram Gajapathi Raj College of Engineering (MVGRCE)",
+  "Maharaja Agrasen Institute of Technology",
+  "Maharaja Surajmal Institute of Technology",
+  "Maharashtra Institute of Technology, Pune",
+  "Mahatma Gandhi Institute for Rural Industrialization (MGIRI)",
+  "Mahatma Gandhi Institute of Technology (MGIT)",
+  "Mahendra Engineering College",
+  "Mailam Engineering College",
+  "Maine South High School",
+  "Maitreyi College, University of Delhi",
+  "Majhighariani Institute Of technology & Science (MITS)",
+  "Malaviya National Institute of Technology Jaipur",
+  "Malineni Lakshmaiah Women's Engineering College",
+  "Malla Reddy College of Engineering Technology",
+  "Malla Reddy Engineering College (MREC)",
+  "Malla Reddy Institute Of Engineering And Technology (MRIET)",
+  "Malnad College of Engineering",
+  "Malvern Preparatory School",
+  "Manakula Vinayagar Institute of Techology",
+  "Manalapan High School",
+  "Manav Rachna International",
+  "Manchester Metropolitan University",
+  "Manhattan College",
+  "Manhattan High School",
+  "Manipal Institute of Technology",
+  "Manipal University",
+  "Manipal University, Jaipur",
+  "Manor College",
+  "Mar Athanasius College of Engineering",
+  "Marc Garneau Collegiate Institute",
+  "Marcellus High School",
+  "Mariana Bracetti Academy Charter School",
+  "Marianopolis College",
+  "Marist College",
+  "Maritime Academy Charter School (MACHS)",
+  "Markham District High School",
+  "Markville Secondary School",
+  "Marlboro High School",
+  "Marquette University",
+  "Marshall High School",
+  "Martin Luther King High School",
+  "Marymount University",
+  "Masaryk University",
+  "Massachusetts Institute of Technology",
+  "Mastery Charter School - Hardy Williams Academy",
+  "Mastery Charter School - Thomas Campus",
+  "Mastery Charter School at Lenfest Campus",
+  "Mastery Charter School at Pickett Campus",
+  "Mastery Charter School at Shoemaker Campus",
+  "Mata Gujri College",
+  "Mater Academy High School",
+  "Math, Civics and Sciences Charter School - Philadelphia",
+  "Mathematics, Science, and Technology Community Charter School (MaST)",
+  "Matrusri Engineering College,  Hyderabad",
+  "Maulana Abul Kalam Azad University of Technology",
+  "Maulana Azad National Institute of Technology",
+  "Maulana Azad National Institute of Technology Bhopal",
+  "Maumee Valley Country Day School",
+  "McGill University",
+  "McMaster University",
+  "Medgar Evers College, CUNY",
+  "Medical University of Silesia",
+  "Meerut Institute of Engineering and Technology (MIET)",
+  "Menlo School",
+  "Mepco Schlenk Engineering College",
+  "Merced College",
+  "Mercer County Community College",
+  "Mercer University",
+  "Meredith College",
+  "Messiah College",
+  "Metas Adventist School",
+  "Metropolia University of Applied Sciences",
+  "Metropolitan State University",
+  "Metuchen High School",
+  "Mewar University Chittorgarh",
+  "Miami Dade College",
+  "Miami Lakes Educational Center",
+  "Miami University",
+  "Michigan State University",
+  "Michigan Technological University",
+  "Microsoft School of the Future High School",
+  "Middle Tennessee State University",
+  "Middlebury College",
+  "Middlesex County Academy",
+  "Middlesex County Academy For Allied Health And Biomedical Sciences",
+  "Middlesex County Academy for Science, Mathematics & Engineering Technologies",
+  "Middlesex County College",
+  "Middlesex University",
+  "Middleton High School",
+  "Middletown High School South",
+  "Midwood",
+  "Miles College",
+  "Millburn High School",
+  "Millburn Middle School",
+  "Millville Senior High School",
+  "Milwaukee School of Engineering",
+  "Minerva University",
+  "Minnesota State University, Mankato",
+  "Misrimal Navajee Munoth Jain Engineering College",
+  "Mission College Boulevard",
+  "Mission San Jose High School",
+  "Mississippi State University",
+  "Mississippi University for Women",
+  "Missouri State University",
+  "Missouri University of Science and Technology",
+  "Model Institute of Engineering and Technology (MIET)",
+  "Modern Engineering and Management Studies",
+  "Mody University",
+  "Mohammed V University",
+  "Molloy College",
+  "Monmouth College",
+  "Monmouth University",
+  "Monroe Community College",
+  "Monroe Township High School",
+  "Monta Vista High School",
+  "Montana State University",
+  "Montclair High School",
+  "Montclair State University",
+  "Montgomery Blair High School",
+  "Montgomery College",
+  "Montgomery County Community College - Central Campus (Blue Bell)",
+  "Montgomery County Community College - West Campus (Pottstown)",
+  "Montgomery High School",
+  "Montville Township High School",
+  "Moore College of Art and Design",
+  "Moore Middle School",
+  "Moorestown High School",
+  "Moraine Valley Community College",
+  "Morehouse College",
+  "Morgan State University",
+  "Morris County School of Technology",
+  "Morris Hills High School",
+  "Morton College",
+  "Moscow Institute of Physics and Technology",
+  "Moscrop Secondary School",
+  "Motilal Nehru National Institute of Technology Allahabad",
+  "Motivation High School (formerly John Bartram High School)",
+  "Mount Holyoke College",
+  "Mountain Lakes High School",
+  "Mountain View High School",
+  "Mt. San Antonio College",
+  "Muhlenberg college",
+  "Multi-Cultural Academy Charter School",
+  "Murrell Dobbins Technical High School",
+  "Muthoot Institute of Technology & Science",
+  "Muzaffarpur Institute of Technology",
+  "NIFT-TEA College of Knitwear Fashion",
+  "NIIT University",
+  "NITK Science & Technology Entrepreneurs' Park (NITK-STEP)",
+  "NMAM Institute of Technology",
+  "NRI Institute of information Science and Technology (NIIST)",
+  "NSS College of Engineering",
+  "Nagaland University, Dimapur Campus",
+  "Nalla Malla Reddy Engineering College, Ghatkesar",
+  "Nanyang Technological University",
+  "Narsee Monjee College of Commerce and Economics",
+  "Narsihma Reddy Engineering College",
+  "Nashua High School South",
+  "National Engineering College",
+  "National Institute of Engineering, Mysore",
+  "National Institute of Science and Technology, Odisha",
+  "National Institute of Technology, Agartala",
+  "National Institute of Technology, Calicut",
+  "National Institute of Technology, Delhi",
+  "National Institute of Technology, Durgapur",
+  "National Institute of Technology, Goa",
+  "National Institute of Technology, Hamirpur",
+  "National Institute of Technology, Jamshedpur",
+  "National Institute of Technology, Karnataka",
+  "National Institute of Technology, Kurukshetra",
+  "National Institute of Technology, Patna",
+  "National Institute of Technology, Raipur",
+  "National Institute of Technology, Rourkela",
+  "National Institute of Technology, Silchar",
+  "National Institute of Technology, Srinagar",
+  "National Institute of Technology, Surat",
+  "National Institute of Technology, Tiruchirappalli",
+  "National Institute of Technology, Trichy",
+  "National Institute of Technology, Uttarakhand",
+  "National Institute of Technology, Warangal",
+  "National Research University Higher School Of Economics",
+  "National University of Singapore",
+  "Neotia Institute Of Technology Management and Science (NITMAS)",
+  "Netaji Subhas Institute of Technology",
+  "Netaji Subhash Engineering College",
+  "Neumann University",
+  "New Albany High School",
+  "New Foundations Charter School - Philadelphia",
+  "New Horizon College of Engineering",
+  "New Jersey City University",
+  "New Jersey Institute of Technology",
+  "New Providence High School",
+  "New River Community College",
+  "New York City College of Technology, CUNY",
+  "New York Institute of Technology",
+  "New York University",
+  "New York University Abu Dhabi",
+  "Newark Charter High School",
+  "Newark Charter Junior/Senior High School",
+  "Newcastle University",
+  "Newton South High School",
+  "Niagara College",
+  "Nipissing University",
+  "Nirma University",
+  "Nitte Meenakshi Institute of Technology",
+  "Nizam College of Engineering Technology",
+  "Noakhali Science and Technology University",
+  "Noida Institute of Engineering and Technology",
+  "Noor-ul-Iman",
+  "Norco College",
+  "Norfolk State University",
+  "North American University",
+  "North Andover High School",
+  "North Brunswick Township High School",
+  "North Carolina Agricultural and Technical (A&T) State University",
+  "North Carolina Central University",
+  "North Carolina School of Science and Mathematics",
+  "North Carolina State University",
+  "North Dakota State University",
+  "North Hunterdon High School",
+  "North Park Secondary School",
+  "North Penn High School",
+  "North Shore Community College",
+  "Northeast High School - Philadelphia",
+  "Northeastern University",
+  "Northern Alberta Institute of Technology (NAIT)",
+  "Northern Arizona University",
+  "Northern Illinois University",
+  "Northern Kentucky University",
+  "Northern Michigan University",
+  "Northern Secondary School",
+  "Northern Virginia Community College",
+  "Northumbria University",
+  "Northview High School",
+  "Northwest Missouri State University",
+  "Northwest Parkway High School",
+  "Northwest Vista College",
+  "Northwestern Oklahoma State University",
+  "Northwestern University",
+  "Northwood Academy/Arts School",
+  "Nottingham Trent University",
+  "Novi High School",
+  "Oakland Community College",
+  "Oakland University",
+  "Obafemi Awolowo University Ile-Ife",
+  "Oberlin College",
+  "Ocean City High School",
+  "Ocean County College",
+  "Oglethorpe University",
+  "Ohio Christian University",
+  "Ohio University",
+  "Okemos High School",
+  "Oklahoma State University",
+  "Old Dominion University",
+  "Old Westbury, SUNY",
+  "Olney High School",
+  "Onalaska High School",
+  "Onondaga Community College",
+  "Ontario Tech University",
+  "Opolska University of Technology",
+  "Oratary Prep School At Summit",
+  "Oregon State University",
+  "Oriental Group of Institutes",
+  "Orissa Engineering College",
+  "Orleans Technical Institute",
+  "Osbourn Park High School",
+  "Ostbayerische Technische Hochschule Regensburg",
+  "Otterbein University",
+  "Overbrook High School - Philadelphia",
+  "Oxford Academy High School",
+  "Oxford Brookes University",
+  "P.D.A. College of Engineering",
+  "PDM College of Engineering",
+  "PES College of Engineering, Mandya",
+  "PES University",
+  "PESIT, Bangalore South Campus",
+  "PGP College of Engineering Technology",
+  "PSG College of Technology,  Coimbatore",
+  "PSG-Science & Technology Entrepreneurial Park (PSG-STEP)",
+  "Pace University",
+  "Pacific University,  Udaipur",
+  "Palo Alto High School",
+  "Palomar College",
+  "Pandit Deendayal Petroleum University",
+  "Panjab University,  SSG Regional Centre",
+  "Parala Maharaja Engineering College,  Berhampur",
+  "Paramount International School",
+  "Park College of Engineering and Technology",
+  "Parkview High School",
+  "Parkway Center City High School",
+  "Parkway West High School",
+  "Parsippany High School",
+  "Parsons School of Design",
+  "Parul Institute of Engineering & Technology",
+  "Pasadena City College",
+  "Pascal English School, Cyprus",
+  "Pathways School Noida",
+  "Patriot High School - Nokesville",
+  "Patriot High School - Riverside",
+  "Paul Robeson High School (formerly John Bartram High School)",
+  "Peirce College",
+  "Penn State Erie, The Behrend College",
+  "Penncrest High School",
+  "Pennington School",
+  "Pennsylvania Academy of the Fine Arts",
+  "Pennsylvania Cyber Charter School",
+  "Pennsylvania Distance Learning Charter School - Online",
+  "Pennsylvania Institute of Technology - Center City Philadelphia",
+  "Pennsylvania Institute of Technology - Media",
+  "Pennsylvania Leadership Charter School - Online",
+  "Pennsylvania Virtual Charter School",
+  "Periyar Maniammai Institute of Science & Technology (PMU)",
+  "Perth Amboy High School",
+  "Perth Amboy Vocational Technical School",
+  "Philadelphia Academy Charter School",
+  "Philadelphia Electrical and Technology Charter School",
+  "Philadelphia High School for Girls",
+  "Philadelphia Performing Arts Charter School (String Theory High School) - Vine Street Campus",
+  "Piedmont High School",
+  "Pierre Elliott Trudeau High School",
+  "Pima Community College",
+  "Pingree School",
+  "Piscataway Township High School",
+  "Pittsburgh Technical College - Philadelphia",
+  "Pittsburgh Technical Institute",
+  "Plaksha University",
+  "Plano East Senior High School",
+  "Plovdiv Medical University",
+  "Point Pleasant Beach High School",
+  "Pokhara University",
+  "Politecnico di Milano",
+  "Polsko-Japońska Akademia Technik Komputerowych",
+  "Pomona College",
+  "Pondicherry Engineering College",
+  "Poolesville High School",
+  "Poornima College of Engineering",
+  "Poornima Group of Institutions",
+  "Poornima Institute Of Engineering And Technology",
+  "Pope John Paul II High School",
+  "Port Credit Secondary School",
+  "Porter-Gaud School",
+  "Portland State University",
+  "Potomac Senior High School",
+  "Potsdam, SUNY",
+  "Poznań University of Technology",
+  "Pranveer Singh Institute of Technology",
+  "Prathyusha Engineering College",
+  "Presidency School, Surat.",
+  "Preston High School",
+  "Preston University",
+  "Princeton Day School",
+  "Princeton High School",
+  "Princeton International School Of Mathematics And Science",
+  "Princeton University",
+  "Priyadarshini College Of Engineering (PEC),  Nagpur",
+  "Proudhadevaraya Institute Of Technology",
+  "Pune Institute of Computer Technology",
+  "Punjab Engineering College (PEC)",
+  "Punjab Institute Of Medical Sciences (PIMS)",
+  "Punjab Institute of Management & Technology",
+  "Punjab Institute of Technology, Rajpura",
+  "Punjab University, Patiala",
+  "Purdue University",
+  "Queen Mary University of London",
+  "Queen's University",
+  "Queens College, CUNY",
+  "Queensborough Community College, CUNY",
+  "R N S Institute of Technology (RNSIT)",
+  "R. R. Institute of Technology",
+  "R.L.Jalappa Institute of technology",
+  "R.V. College Of Engineering",
+  "R.V. College of Engineering (RVCE)",
+  "R.V.R. & J.C. College of Engineering",
+  "RAM-EESH INSTITUTE OF ENGINEERING TECHNOLOGY",
+  "REVA University",
+  "RIMT Institute of Engineering and Technology",
+  "RMK College of Engineering",
+  "RNS Institute of Technology",
+  "RPIIT Technical Campus",
+  "Radharaman Institute of Research Technology (RIRT), Radharaman Group",
+  "Radharaman Institute of Technology & Science (RITS), Bhopal",
+  "Radnor High School",
+  "Raj Kumar Goel Engineering College",
+  "Rajagiri School of Engineering and Technology",
+  "Rajarajeswari College of Engineering (RRCE)",
+  "Rajasthan Institute Of Engineering and Technology",
+  "Rajdhani College of Engineering & Management",
+  "Rajendra Mane College of Engineering and Technology (RMCET)",
+  "Rajiv Gandhi College of Engineering and Technology",
+  "Rajiv Gandhi Institute of Technology (RIT), Kottayam",
+  "Rajiv Gandhi University of Knowledge Technologies (RGUKT), Basar",
+  "Rajkiya Engineering College, Ambedkar Nagar",
+  "Raksha Shakti University",
+  "Ramaiah Institute of Technology",
+  "Ramapo College of New Jersey",
+  "Ramapo High School",
+  "Ramrao Adik Institute of Technology (RAIT), DY Patil University",
+  "Randolph-Macon College",
+  "Rani Laxmi Bai Public School",
+  "Raritan High School",
+  "Raritan Valley Community College",
+  "Rasmussen University",
+  "Ravenscroft School",
+  "Ravenwood High School",
+  "Reach Cyber Charter School",
+  "Red Bank Regional High School",
+  "Reed College",
+  "Regional College For Education Research and Technology, Jaipur",
+  "Regis High School",
+  "Rensselaer Polytechnic Institute",
+  "Rheinisch-Westfälische Technische Hochschule Aachen (RWTH)",
+  "Rhode Island College",
+  "Rhode Island School of Design",
+  "Rhodes College",
+  "Rice University",
+  "Richard Montgomery High School",
+  "Richard Stockton University",
+  "Richardson High School",
+  "Richland College",
+  "Richmond Hill High School",
+  "Rider University",
+  "Ridge High School",
+  "Ridgewood High School",
+  "Riga Technical University",
+  "River Dell High School",
+  "Robbinsville High School",
+  "Robert Gordon University",
+  "Rochester Institute of Technology",
+  "Rock Ridge High School",
+  "Roger Williams University",
+  "Rollins College",
+  "Roosevelt High School",
+  "Rosa Parks Middle School",
+  "Rose-Hulman Institute of Technology",
+  "Rosemont College",
+  "Rowan College at Burlington County - Mount Holly",
+  "Rowan College at Burlington County - Pemberton",
+  "Rowan College at Burlington County - Willingboro",
+  "Rowan College at Gloucester County - Mount Laurel",
+  "Rowan University",
+  "Roxborough High School",
+  "Roxbury High School",
+  "Royal Holloway, University of London",
+  "Rudbecksgymnasiet",
+  "Rungta College of Engineering and Technology, Bhilai",
+  "Rustamji Institute of Technology",
+  "Rutgers Preparatory School",
+  "Rutgers University - Newark",
+  "Rutgers University – Camden",
+  "Rutgers, The State University of New Jersey",
+  "Ryde School",
+  "Rye High School",
+  "Ryerson University",
+  "S A Engineering College",
+  "S G Balekundri Institute of Technology",
+  "SAL Engineering and Technical Institute",
+  "SASTRA University",
+  "SJB Institute of Technology",
+  "SKR Engineering College",
+  "SOAS University of London",
+  "SRM Easwari Engineering College, Chennai",
+  "SRM University",
+  "SRM University, Sonepat",
+  "SS College of Engineering",
+  "SUNY Polytechnic Institute",
+  "SUPINFO International University",
+  "SVS College of Engineering",
+  "Sachdeva Institute of Technology",
+  "Sagar Institute of Science & Technology (SISTec)",
+  "Saginaw Valley State University",
+  "Sahrdaya College of Engineering and Technology",
+  "Sai Vidya Institute of Technology",
+  "Saint Joseph High School",
+  "Saint Joseph's College of Maine",
+  "Saint Joseph's Preparatory School - Philadelphia",
+  "Saint Joseph's University - Philadelphia",
+  "Saint Paul College",
+  "Saint Peter's Preparatory School",
+  "Saint Peter's University",
+  "Salem Community College",
+  "Salem State University",
+  "Sambalpur University Institute of Information Technology (SUIIT)",
+  "Sambhram Institute of Technology",
+  "Samrat Ashok Technological Institute (S.A.I.T)",
+  "Samuel Fels High School - Philadelphia",
+  "San Diego State University",
+  "San Francisco State University",
+  "San Jose State University",
+  "San Marcos High School",
+  "San Marin High School",
+  "San Mateo High School",
+  "Sankofa Freedom Academy Charter School",
+  "Sant Longowal Institute of Engineering and Technology",
+  "Santa Barbara City College",
+  "Santa Clara University",
+  "Santa Margarita Catholic High School",
+  "Santa Rosa Junior College",
+  "Sapthagiri College of Engineering ",
+  "Saratoga High School",
+  "Sardar Patel Institute Of Technology",
+  "Sardar Patel University",
+  "Sardar Vallabhbhai National Institute of Technology, Surat",
+  "Sardar Vallabhbhai Patel Institute of Technology, Vasad",
+  "Sarvajanik College of Engineering & Technology",
+  "Saurashtra University Rajkot",
+  "Savannah State University",
+  "Savitribai Phule Pune University",
+  "School of Engineering and Technology, Mizoram University",
+  "School of Engineering, Cochin University of Science and Technology",
+  "School of Professional Studies, CUNY",
+  "School of Visual Arts, New York",
+  "Science Leadership Academy",
+  "Science and Technology Entrepreneurs Park (STEP), Harcourt",
+  "Science and TechnologyEntrepreneurs Park, Indian Institute of Technology",
+  "Scranton High School",
+  "Seneca College",
+  "Seton Hall University",
+  "Seven Lakes High School",
+  "Seventh Day Adventist High School",
+  "Shaker High School",
+  "Shankersinh Vaghela Bapu Institute of Technology",
+  "Sharda University",
+  "Sheffield Hallam University",
+  "Shelton High School",
+  "Sheridan College",
+  "Sherwood Convent School",
+  "Sherwood High School",
+  "Shiv Nadar University",
+  "Shri Dharmasthala Manjunatheshwara College of Engineering and Technology (SDM)",
+  "Shri Govindram Seksaria Institute of Technology and Science",
+  "Shri Guru Gobind Singhji Institute of Engineering and Technology (SGGS)",
+  "Shri Guru Ram Rai Public School",
+  "Shri Mata Vaishno Devi University(SMVDU)",
+  "Shri Ramswaroop College Of Engineering and Management",
+  "Shri Ramswaroop Memorial Group of Professional Colleges (SRMGPC)",
+  "Shri Sant Gajanan Maharaj College of Engineering, Shegaon (SSGMCE)",
+  "Shri Shankaracharya Technical Campus",
+  "Shri Vaishnav Institute of Technology and Science",
+  "Shri Venkateshwara College of Engineering",
+  "Shridevi Institute of Engineering & Technology",
+  "Shriram Institute for Industrial Research",
+  "Siddaganga Institute Of Technology, Tumakuru",
+  "Siena College",
+  "Sikkim Manipal Institute of Technology",
+  "Silesian University of Technology",
+  "Silicon Institute of Technology",
+  "Siliguri Institute of Technology",
+  "Silver Oak College of Engineering & Technology",
+  "Simmons College",
+  "Simon Fraser University",
+  "Simon Gratz High School",
+  "Simpson College",
+  "Simsbury High School",
+  "Simón Bolívar University",
+  "Sinclair Community College",
+  "Singapore University of Technology and Design",
+  "Sinhgad Institute of Technology",
+  "Sir John A. Macdonald Secondary School",
+  "Sir M Visvesvaraya Institute of Technology (Sir MVIT)",
+  "Sir Padampat Singhania University",
+  "Sitarambhai Naranji Patel Institute of Technology & Research Centre",
+  "Skidmore College",
+  "Slippery Rock University of Pennsylvania",
+  "Slovak University of Technology in Bratislava (STU)",
+  "Smith College",
+  "Society for Development of Composites",
+  "Solebury School",
+  "Sona College of Technology",
+  "Souderton Area High School",
+  "South Brunswick High School",
+  "South Carolina State University",
+  "South Dakota School of Mines and Technology",
+  "South Hills School of Business & Technology",
+  "South Lakes High School",
+  "South Philadelphia High School",
+  "South Texas College",
+  "Southeastern Louisiana University",
+  "Southern Connecticut State University",
+  "Southern Illinois University Carbondale",
+  "Southern Illinois University Edwardsville",
+  "Southern Methodist University",
+  "Southern Oregon University",
+  "Southern University and A&M College",
+  "Southern Utah University",
+  "Southwestern College",
+  "Spelman College",
+  "Spotswood High School",
+  "Spring Arbor University",
+  "Springside Chestnut Hill Academy",
+  "Sree Chitra Thirunal College of Engineering",
+  "Sreenidhi Institute of Science & Technology",
+  "Sri Jayachamarajendra College of Engineering",
+  "Sri Krishna College of Engineering and Technology (SKCET)",
+  "Sri Krishna College of Technology, Coimbatore",
+  "Sri Lanka Institute of Information Technology (SLIIT)",
+  "Sri Manakula Vinayagar Engineering",
+  "Sri Ramakrishna Engineering College (SREC)",
+  "Sri Revana Siddeshwara Institute of Technology",
+  "Sri Siddhartha Institute of Technology",
+  "Sri Sivasubramaniya Nadar College of Engineering",
+  "Sri Venkateshwara College of Engineering",
+  "Sri Vishnu Educational Society",
+  "Srinivas Institute of Technology (SIT)",
+  "St Brendan High School",
+  "St Edwards University",
+  "St Joseph Engineering College",
+  "St Mary's CE High School – Cheshunt",
+  "St Mary's Catholic High School – Croydon",
+  "St Paul's Catholic College – Sunbury-on-Thames",
+  "St. Charles Borromeo Seminary",
+  "St. Cloud State University",
+  "St. David Catholic Secondary School",
+  "St. John's University, New York",
+  "St. Joseph's College of Engineering and Technology, Palai",
+  "St. Mark's School, Hong Kong",
+  "St. Mary's Convent School",
+  "St. Mary's Ryken High School",
+  "St. Michael College of Engineering & Technology",
+  "St. Peter's Institute of Higher Education and Research",
+  "St. Pious X Degree & PG College for women",
+  "St. Raymond High School for Boys And Girls",
+  "St. Theresa of Lisieux Catholic High School",
+  "St. Xavier's Senior Secondary School, Jaipur",
+  "St.Martin's Engineering College",
+  "Stanford University",
+  "Stanley College of Engineering and Technology for Women",
+  "Star Technical Institute",
+  "Startup Incubation and Innovation Centre, IIT Kanpur",
+  "Staten Island Technical High School",
+  "Steinert High School",
+  "Stephen F. Austin State University",
+  "Stetson University",
+  "Stevens Institute of Technology",
+  "Stevenson University",
+  "Stockholm University",
+  "Stockton University",
+  "Stonehill College",
+  "Stonewall Jackson High School - Manassas",
+  "Stonewall Jackson High School - Quicksburg",
+  "Stony Brook University, SUNY",
+  "Strawberry Mansion High School",
+  "Strayer University - Bensalem",
+  "Strayer University - Philadelphia Center City",
+  "Stuyvesant High School",
+  "Sulphur High School",
+  "Susq-Cyber Charter School",
+  "Susquehanna University",
+  "Sussex County Community College",
+  "Suyash Institute of Information Technology",
+  "Swami Keshvanand Institute of Technology,  Management & Gramothan (SKIT)",
+  "Swansea University",
+  "Swarthmore College",
+  "Syed Ammal Engineering College",
+  "Symbiosis International University",
+  "Synergy Institute of Engineering and Technology",
+  "Syracuse University",
+  "T K M College of Engineering",
+  "TECH Freire Charter High School",
+  "THDC Institute of Hydropower Engineering and Technology, Tehri",
+  "TU/e Technische Universiteit Eindhoven University of Technology",
+  "Tacoma Community College",
+  "Tacony Academy Charter School",
+  "Tadeusz Kościuszko University of Technology",
+  "Tallinn University",
+  "Tallinn University of Technology",
+  "Talmudical Yeshiva of Philadelphia",
+  "Tamil Nadu Agricultural University (TNAU)",
+  "Tamilnadu College of Engineering",
+  "Tampere University of Applied Sciences",
+  "Tampere University of Technology",
+  "Tarleton State University",
+  "Technische Universität München",
+  "Techno India College of Technology",
+  "Techno India University",
+  "Tecnológico de Estudio Superiores de Ixtapaluca",
+  "Tecnológico de Estudios Superiores de Ecatepec",
+  "Tecnológico de Estudios Superiores de Jilotepec",
+  "Teesside University",
+  "Temple University",
+  "Temple University - Ambler",
+  "Temple University - Harrisburg",
+  "Temple University - Health Sciences Campus",
+  "Temple University - Rome",
+  "Temple University - Tokyo",
+  "Tenafly High School",
+  "Tennessee State University",
+  "Texas A&M University",
+  "Texas A&M University – Central Texas",
+  "Texas A&M University – Corpus Christi",
+  "Texas A&M University – Kingsville",
+  "Texas Christian University",
+  "Texas Southern University",
+  "Texas Southmost College",
+  "Texas State University",
+  "Texas Tech University",
+  "Tezpur University",
+  "Thadomal Shahani Engineering College",
+  "Thakur College of Engineering and Technology",
+  "Thanthai Periyar Government Institute of Technology",
+  "Thapar Institute of Engineering and Technology",
+  "The Arts Academy at Benjamin Rush",
+  "The British University In Egypt",
+  "The Bronx High School of Science",
+  "The City College of New York, CUNY",
+  "The College at Brockport, SUNY",
+  "The College of New Jersey",
+  "The College of Saint Rose",
+  "The Curtis Institute of Music",
+  "The Federal University of Technology,  Akure",
+  "The George Washington University",
+  "The Governor's School @ Innovation Park",
+  "The Harker School",
+  "The Hill School",
+  "The Katholieke Universiteit Leuven",
+  "The LNM Institute of Information Technology",
+  "The Lawrenceville School",
+  "The Maharaja Sayajirao University of Baroda",
+  "The Mount Tabor Training College",
+  "The Ohio State University",
+  "The Open University",
+  "The Oxford College of Engineering",
+  "The Pennsylvania State University",
+  "The Pennsylvania State University – Abington Campus",
+  "The Pennsylvania State University – Berks",
+  "The Pennsylvania State University – Brandywine",
+  "The Pennsylvania State University – Harrisburg",
+  "The Pennsylvania State University – York Campus",
+  "The Roxbury Latin School",
+  "The SRM University",
+  "The Savannah College of Art and Design",
+  "The Technical University of Denmark",
+  "The Technische Universität Berlin",
+  "The University of Aberdeen",
+  "The University of Akron",
+  "The University of Alabama",
+  "The University of Alabama at Birmingham",
+  "The University of Alberta",
+  "The University of Applied Sciences Upper Austria",
+  "The University of Arizona",
+  "The University of Arkansas",
+  "The University of Bath",
+  "The University of Bedfordshire",
+  "The University of Birmingham",
+  "The University of Bolton",
+  "The University of Bonn",
+  "The University of Bristol",
+  "The University of British Columbia",
+  "The University of Calgary",
+  "The University of Calicut",
+  "The University of California, Berkeley",
+  "The University of California, Davis",
+  "The University of California, Irvine",
+  "The University of California, Los Angeles",
+  "The University of California, Merced",
+  "The University of California, Riverside",
+  "The University of California, San Diego",
+  "The University of California, Santa Barbara",
+  "The University of California, Santa Cruz",
+  "The University of Cambridge",
+  "The University of Chicago",
+  "The University of Colorado Boulder",
+  "The University of Colorado Colorado Springs",
+  "The University of Connecticut",
+  "The University of Dallas",
+  "The University of Delaware",
+  "The University of Denver",
+  "The University of Derby",
+  "The University of Dundee",
+  "The University of Edinburgh",
+  "The University of Essex",
+  "The University of Evansville",
+  "The University of Exeter",
+  "The University of Falmouth",
+  "The University of Gdańsk",
+  "The University of Georgia",
+  "The University of Glasgow",
+  "The University of Groningen",
+  "The University of Guelph",
+  "The University of Houston",
+  "The University of Houston – Clear Lake",
+  "The University of Houston – Downtown",
+  "The University of Huddersfield",
+  "The University of Idaho",
+  "The University of Illinois at Chicago",
+  "The University of Illinois at Urbana-Champaign",
+  "The University of Information Technology and Management in Rzeszow",
+  "The University of Iowa",
+  "The University of Kansas",
+  "The University of Kent",
+  "The University of Kentucky",
+  "The University of La Verne",
+  "The University of Leeds",
+  "The University of Leicester",
+  "The University of Lincoln",
+  "The University of Liverpool",
+  "The University of Ljubljana",
+  "The University of Louisiana at Lafayette",
+  "The University of Louisiana at Monroe",
+  "The University of Louisville",
+  "The University of Manchester",
+  "The University of Manitoba",
+  "The University of Maryland, Baltimore County",
+  "The University of Maryland, College Park",
+  "The University of Massachusetts Amherst",
+  "The University of Massachusetts Boston",
+  "The University of Massachusetts Dartmouth",
+  "The University of Massachusetts Lowell",
+  "The University of Miami",
+  "The University of Michigan",
+  "The University of Michigan-Dearborn",
+  "The University of Michigan-Flint",
+  "The University of Minnesota",
+  "The University of Mississippi",
+  "The University of Missouri",
+  "The University of Missouri-Kansas City",
+  "The University of Missouri-St. Louis",
+  "The University of Málaga",
+  "The University of Nebraska-Lincoln",
+  "The University of New Brunswick",
+  "The University of New Hampshire",
+  "The University of New Haven",
+  "The University of North Carolina at Chapel Hill",
+  "The University of North Carolina at Charlotte",
+  "The University of North Carolina at Greensboro",
+  "The University of North Texas",
+  "The University of Northampton",
+  "The University of Notre Dame",
+  "The University of Nottingham",
+  "The University of Oklahoma",
+  "The University of Oregon",
+  "The University of Ottawa",
+  "The University of Oulu",
+  "The University of Oxford",
+  "The University of Pennsylvania",
+  "The University of Petroleum and Energy Studies",
+  "The University of Phoenix",
+  "The University of Pittsburgh",
+  "The University of Portland",
+  "The University of Portsmouth",
+  "The University of Puerto Rico, Mayagüez Campus",
+  "The University of Puerto Rico, Río Piedras Campus",
+  "The University of Richmond",
+  "The University of Rochester",
+  "The University of Salford",
+  "The University of San Francisco",
+  "The University of Sharjah",
+  "The University of Sheffield",
+  "The University of Silesia in Katowice",
+  "The University of South Carolina",
+  "The University of Southampton",
+  "The University of Southern California",
+  "The University of Southern Denmark",
+  "The University of St Andrews",
+  "The University of St. Gallen",
+  "The University of St. Thomas",
+  "The University of Stirling",
+  "The University of Strathclyde",
+  "The University of Stuttgart",
+  "The University of Surrey",
+  "The University of Sussex",
+  "The University of Tampa",
+  "The University of Tennessee",
+  "The University of Texas Rio Grande Valley",
+  "The University of Texas at Arlington",
+  "The University of Texas at Austin",
+  "The University of Texas at Dallas",
+  "The University of Texas at El Paso",
+  "The University of Texas at San Antonio",
+  "The University of Texas at Tyler",
+  "The University of Texas of the Permian Basin",
+  "The University of Texas – Pan American",
+  "The University of Toledo",
+  "The University of Toronto",
+  "The University of Toronto Mississauga",
+  "The University of Toronto Scarborough",
+  "The University of Tulsa",
+  "The University of Utah",
+  "The University of Vermont",
+  "The University of Victoria",
+  "The University of Virginia",
+  "The University of Warsaw",
+  "The University of Warwick",
+  "The University of Washington",
+  "The University of Washington Bothell",
+  "The University of Waterloo",
+  "The University of West Georgia",
+  "The University of Western Ontario",
+  "The University of Westminster",
+  "The University of Windsor",
+  "The University of Wisconsin-Eau Claire",
+  "The University of Wisconsin-Green Bay",
+  "The University of Wisconsin-La Crosse",
+  "The University of Wisconsin-Madison",
+  "The University of Wisconsin-Milwaukee",
+  "The University of Wisconsin-Oshkosh",
+  "The University of Wisconsin-Parkside",
+  "The University of Wisconsin-Platteville",
+  "The University of Wisconsin-River Falls",
+  "The University of Wisconsin-Stevens Point",
+  "The University of Wisconsin-Stout",
+  "The University of Wisconsin-Superior",
+  "The University of Wisconsin-Whitewater",
+  "The University of Wolverhampton",
+  "The University of Wrocław",
+  "The University of York",
+  "The University of Zagreb",
+  "The University of the District of Columbia",
+  "The University of the Pacific",
+  "The University of the South - Sewanee",
+  "The University of the West Indies",
+  "The Université de Sherbrooke",
+  "The Workshop School - Philadelphia",
+  "Thiagarajar College of Engineering (TCE), Madurai",
+  "Thomas A. Edison High School - Philadelphia",
+  "Thomas Edison State College",
+  "Thomas Jefferson High School for Science and Technology",
+  "Thomas Jefferson University - East Falls (formerly Philadelphia University)",
+  "Thomas Jefferson University - Philadelphia Center City",
+  "Thomas Nelson Community College",
+  "Thomas S. Wootton High School",
+  "Thompson Institute - Philadelphia",
+  "Tiruchirappalli Regional Engineering College Science Technology",
+  "Tongji University",
+  "Towson High School",
+  "Towson University",
+  "Trent University",
+  "Trident Academy of Technology",
+  "Trinity College",
+  "Trinity International University",
+  "Trinity Valley School",
+  "Troy Athens High School",
+  "Troy High School",
+  "Troy University",
+  "Truman State University",
+  "Tshwane University of Technology",
+  "Tufts University",
+  "Tulane University",
+  "Tunis El Manar University",
+  "Turner Fenton Secondary School",
+  "UNAM FES Aragón",
+  "Ulster University",
+  "Union County College",
+  "Union County Magnet High School",
+  "Union County Vocational-Technical Schools",
+  "Union University",
+  "Unionville High School",
+  "United College of Engineering and Research",
+  "United Institute of Technology",
+  "Universidad Autónoma Metropolitana",
+  "Universidad Autónoma de Baja California (UABC), Tijuana",
+  "Universidad Autónoma de Coahuila",
+  "Universidad Autónoma de Madrid",
+  "Universidad Autónoma de Nuevo León",
+  "Universidad Autónoma de San Luis Potosí",
+  "Universidad Autónoma de Tlaxcala",
+  "Universidad Autónoma del Estado de Morelos",
+  "Universidad Autónoma del Estado de México",
+  "Universidad Autónoma del Perú",
+  "Universidad Centro de Estudios Cortazar",
+  "Universidad Iberoamericana",
+  "Universidad Interamericana de Puerto Rico",
+  "Universidad Nacional Autónoma de México",
+  "Universidad Panamericana",
+  "Universidad Politécnica de Guanajuato",
+  "Universidad Politécnica de Querétaro",
+  "Universidad TecMilenio",
+  "Universidad Tecnológica Nacional",
+  "Universidad Tecnológica de México",
+  "Universidad Tecnológica de Puebla",
+  "Universidad Tecnológica de Torreón",
+  "Universidad Veracruzana",
+  "Universidad de Guadalajara",
+  "Universidad de Guanajuato",
+  "Universidad de La Laguna",
+  "Universidad de La Salle Bajío",
+  "Universidad de Monterrey",
+  "Universidad del Desarrollo",
+  "Universidad del Valle de México",
+  "Universidad en Línea, Mexico",
+  "Universitat Autònoma de Barcelona, UAB",
+  "Universitat Oberta de Catalunya, UOC",
+  "Universitat Politècnica de Catalunya",
+  "Universitat Politècnica de Catalunya, UPC",
+  "Universitat Pompeu Fabra",
+  "Universitat de Barcelona",
+  "Universitatea Politehnica Timişoara",
+  "University Campus Suffolk",
+  "University College London",
+  "University College of Engineering and Technology,  Bikaner",
+  "University Institute of Engineering and Technology CSJMU, Kanpur",
+  "University Institute of Information Technology, Shimla",
+  "University Institute of Technology, Burdwan",
+  "University Institute of Technology, RGPV",
+  "University School of Information, Communication and Technology",
+  "University Visvesvaraya College of Engineering (UVCE)",
+  "University at Albany, SUNY",
+  "University at Binghamton, SUNY",
+  "University at Buffalo, SUNY",
+  "University at New Paltz, SUNY",
+  "University at Oneonta, SUNY",
+  "University at Orange, SUNY",
+  "University at Oswego, SUNY",
+  "University at Plattsburgh, SUNY",
+  "University of Basel",
+  "University of Białystok",
+  "University of Cape Coast, Ghana",
+  "University of Cincinnati",
+  "University of Cincinnati Clermont College",
+  "University of Duisburg-Essen",
+  "University of Gothenburg",
+  "University of Helsinki",
+  "University of Hull",
+  "University of London",
+  "University of Mary Washington",
+  "University of Maryland University College",
+  "University of North America",
+  "University of North Georgia",
+  "University of Petroleum and Energy Studies (UPES), Dehradun",
+  "University of Pikeville",
+  "University of Queensland",
+  "University of Regina",
+  "University of Roehampton",
+  "University of Saskatchewan",
+  "University of Science and Technology Houari Boumediene",
+  "University of Southampton",
+  "University of Southern Indiana",
+  "University of Sunderland",
+  "University of Tartu",
+  "University of Technology, Jamaica",
+  "University of Trento",
+  "University of Udaipur",
+  "University of Valley Forge",
+  "University of Washington Tacoma",
+  "University of the Arts - Philadelphia",
+  "University of the People",
+  "University of the Sciences in Philadelphia",
+  "Universität Regensburg",
+  "Universität Zürich",
+  "Université de Bordeaux",
+  "Université de Mons",
+  "Université du Québec à Montréal",
+  "Upper Canada College",
+  "Upper Darby High School",
+  "Upper Iowa University",
+  "Upper Moreland High School",
+  "Urbana High School",
+  "Ursinus College",
+  "Utah State University",
+  "Utica College",
+  "Utkal University",
+  "Uttaranchal Institute of Technology",
+  "VIA University College",
+  "VNS Group of Institutions",
+  "Vadodara Institute of Engineering",
+  "Valencia College",
+  "Valley Christian High School",
+  "Valley High School",
+  "Vallurupalli Nageswara Rao Vignana Jyothi Institute of Engg. Technology (VNRVJIET)",
+  "Vanderbilt University",
+  "Vanier College",
+  "Vasavi College Of Engineering",
+  "Vassar College",
+  "Veer Narmad South Gujarat University",
+  "Veer Surendra Sai University of Technology",
+  "Veer Surendra Sai University of Technology, Burla",
+  "Vel Tech Multi Tech Dr.Rangarajan Dr.Sakunthala Engineering College",
+  "Vel Tech Rangarajan Dr.Sagunthala R&D Institute of Science and Technology",
+  "Velammal College of Engineering and Technology",
+  "Velammal Institute of Technology",
+  "Vellore Institute of Technology",
+  "Vellore Institute of Technology, Chennai",
+  "Vemana Institute Of Technology",
+  "Veterans Memorial Early College High School",
+  "Victoria Park Collegiate Institute",
+  "Vidya College of Engineering",
+  "Vidyakunj International School",
+  "Vidyavardhaka College of Engineering",
+  "Vignan Institute of Technology and Science",
+  "Vikas College of Engineering & Technology, Vijayawada",
+  "Villanova University",
+  "Villgro Innovations Foundation IITM Research Park",
+  "Vinayaka Mission's Kirupananda Variyar Engineering College",
+  "Vincennes University",
+  "Vincent Massey Secondary School",
+  "Virginia Commonwealth University",
+  "Virginia State University",
+  "Virginia Tech",
+  "Virginia Union University",
+  "Virginia University of Lynchburg",
+  "Virtual High School @ PWCS",
+  "Vishwakarma Government Engineering College",
+  "Vishwakarma Institute of Technology",
+  "Visvesvaraya National Institute of Technology",
+  "Visvesvaraya Technological University",
+  "Vivekanand Education Society's Institute of Technology (VESIT)",
+  "Vivekanand Institute of Technology & Sciences",
+  "Vivekananda College for BCA",
+  "Vivekananda Institute of Biotechnology",
+  "Vivekananda Institute of Technology",
+  "Vizag Institute of Technology",
+  "Vrije Universiteit Amsterdam",
+  "Wake Forest University",
+  "Wake Technical Community College",
+  "Walchand College of Engineering",
+  "Walnut Hill College",
+  "Walt Whitman High School",
+  "Walter Biddle Saul High School",
+  "Ward Melville High School",
+  "Wardlaw + Hartridge School",
+  "Warren County Technical High School",
+  "Warsaw School of Economics",
+  "Warsaw University of Technology",
+  "Wartburg College",
+  "Washington State University",
+  "Washington Township High School",
+  "Washington University in St. Louis",
+  "Washington and Lee University",
+  "Waterloo Collegiate Institute",
+  "Waunakee High School",
+  "Wayne State University",
+  "Webb Bridge Middle School",
+  "Wellesley College",
+  "Wellington C. Mepham Highschool",
+  "Wells College",
+  "Wentworth Institute of Technology",
+  "Wesleyan University",
+  "West Chester University",
+  "West Essex Regional High School",
+  "West Morris Mendham High School",
+  "West Philadelphia High School",
+  "West Potomac High School",
+  "West Scranton High School",
+  "West Windsor-Plainsboro High School North",
+  "West Windsor-Plainsboro High School South",
+  "Westdale Secondary School",
+  "Western Carolina University",
+  "Western Connecticut State University",
+  "Western Governors University",
+  "Western Kentucky University",
+  "Western Michigan University",
+  "Western New England University",
+  "Western Technical College",
+  "Western University",
+  "Western Washington University",
+  "Westfield High School",
+  "Westminster College",
+  "Westminster School",
+  "Westwood High School",
+  "Whitefish Bay High School",
+  "Whitworth University",
+  "Wichita State University",
+  "Widener University",
+  "Wilbert Tucker Woodson High School",
+  "Wilfrid Laurier University",
+  "Wilkes University",
+  "William & Mary",
+  "William L. Sayre High School",
+  "William Lyon Mackenzie Collegiate Institute",
+  "William Paterson University",
+  "William W. Bodine High School",
+  "Williams College",
+  "Williamson Free School of Mechanical Trades",
+  "Wilmington University",
+  "Wiltshire College",
+  "Winona State University",
+  "Winston Churchill High School",
+  "Winthrop University",
+  "Woodbridge High School - Bridgeville",
+  "Woodbridge High School - Irvine",
+  "Woodbridge High School - London",
+  "Woodbridge High School - Woodbridge, NJ",
+  "Woodbridge High School - Woodbridge, ON",
+  "Woodbridge High School - Woodbridge, VA",
+  "Worcester Polytechnic Institute",
+  "Worcester State University",
+  "World Communications Charter School",
+  "Wright State University",
+  "Wrocław University of Economics",
+  "Wrocław University of Technology",
+  "Wuhan University",
+  "Wyższa Szkoła Biznesu – National-Louis University",
+  "Xavier Institute of Management Entrepreneurship Development (XIME)",
+  "Xavier Research Foundation Loyola Centre for Research and Development St Xavier's College",
+  "Xavier University",
+  "Yale University",
+  "Yale-NUS College",
+  "Yeshiva University",
+  "York College of Pennsylvania",
+  "York College, CUNY",
+  "York University",
+  "Youngstown State University",
+  "YouthBuild Philadelphia Charter School",
+  "Zakir Hussain College of Engineering and Technology, AMU",
+  "Zespół Szkół Nr.2 im. Jana Pawła II w Miechowie",
+  "Zespół Szkół im. Jana Pawła II w Niepołomicach",
+  "Zespół Szkół nr 1 im. Jana Pawła II w Przysusze",
+  "Zespół Szkół Łączności, Monte Cassino 31",
+  "Zespół szkół nr 1 im. Stanisława Staszica w Bochni",
+  "ecole centrale marseille",
+  "vardhaman college of engineering",
+  "École Centrale Paris",
+  "École Polytechnique de Montréal",
+  "École de technologie supérieure",
+  "École nationale supérieure d’électronique, informatique, télécommunications, mathématique et mécanique de Bordeaux (ENSEIRB-MATMECA)",
+] as const;
+
+export const SHIRT_SIZES = ["SM", "MD", "LG", "XL", "XXL"] as const;
+
+export const SPONSOR_TIERS = ["gold", "silver", "bronze", "other"] as const;
+
+export const US_STATES = [
+  "Florida",
+  "Georgia",
+  "California",
+  "Massachusetts",
+  "New York",
+  "Alabama",
+  "Alaska",
+  "Arizona",
+  "Arkansas",
+  "Colorado",
+  "Connecticut",
+  "Delaware",
+  "Hawaii",
+  "Idaho",
+  "Illinois",
+  "Indiana",
+  "Iowa",
+  "Kansas",
+  "Kentucky",
+  "Louisiana",
+  "Maine",
+  "Maryland",
+  "Michigan",
+  "Minnesota",
+  "Mississippi",
+  "Missouri",
+  "Montana",
+  "Nebraska",
+  "Nevada",
+  "New Hampshire",
+  "New Jersey",
+  "New Mexico",
+  "North Carolina",
+  "North Dakota",
+  "Ohio",
+  "Oklahoma",
+  "Oregon",
+  "Pennsylvania",
+  "Rhode Island",
+  "South Carolina",
+  "South Dakota",
+  "Tennessee",
+  "Texas",
+  "Utah",
+  "Vermont",
+  "Virginia",
+  "Washington",
+  "West Virginia",
+  "Wisconsin",
+  "Wyoming",
+] as const;
