--- packages/consts/src/majors.ts
@@ -1,203 +1,0 @@
-export const MAJORS = [
-  "Computer Science",
-  "Computer Engineering",
-  "Electrical Engineering",
-  "Information Technology (IT)",
-  "Software Engineering",
-  "Aeronautics",
-  "Agricultural Business & Technology",
-  "Agriculture",
-  "Agricultural Economics",
-  "Agricultural Operations",
-  "Animal Sciences",
-  "Construction Engineering Technology",
-  "Construction Management",
-  "Fish and Wildlands Science and Management",
-  "Food Science",
-  "Animation",
-  "Art",
-  "Art History",
-  "Cinematography",
-  "Dance",
-  "Digital Arts",
-  "Drama",
-  "Fashion Design",
-  "Film",
-  "Game Design",
-  "Graphic Design",
-  "Industrial Design",
-  "Illustration",
-  "Interior Design",
-  "Multimedia",
-  "Music",
-  "Music Management",
-  "Music Performance",
-  "Photography",
-  "Studio Art",
-  "Visual and Performing Arts",
-  "Visual Design",
-  "Accounting",
-  "Actuarial Science",
-  "Advertising",
-  "Business",
-  "Business Administration",
-  "Business, Management, and Related Support Services",
-  "Econometrics",
-  "Economics",
-  "Entrepreneurship",
-  "Fashion Merchandising",
-  "Finance",
-  "Human Resources Management",
-  "International Business",
-  "Insurance",
-  "Management",
-  "Management Information Systems",
-  "Management Science",
-  "Managerial Economics",
-  "Marketing",
-  "Operations Management",
-  "Organizational Behavior",
-  "Real Estate",
-  "Sales",
-  "Supply Chain Management",
-  "Early Childhood Education",
-  "Education",
-  "Education (Other)",
-  "Elementary Education",
-  "Health and Physical Education",
-  "Language Arts Teacher Education",
-  "Mathematics Teacher Education",
-  "Middle School Education",
-  "Music Teacher Education",
-  "Physical Education Teaching and Coaching",
-  "Secondary Education",
-  "Special Education",
-  "Teacher Education",
-  "Athletic Training",
-  "Audiology",
-  "Clinical Laboratory Science",
-  "Cognitive Science",
-  "Communication Sciences",
-  "Community Health",
-  "Community Health Services",
-  "Dental Hygiene",
-  "Health and Wellness",
-  "Health Care Administration",
-  "Health Information",
-  "Health/Medical Preparatory Programs",
-  "Health Professions and Related Clinical Sciences",
-  "Health Services",
-  "Hospital and Health Care Facilities Administration",
-  "Medical Radiologic Technology",
-  "Nursing",
-  "Nursing Research",
-  "Nursing Science",
-  "Nutrition",
-  "Nutrition Sciences",
-  "Public Health",
-  "Public Health Education",
-  "Radiologic Technology",
-  "Respiratory Care Therapy",
-  "Speech-Language Pathology",
-  "Applied Communication",
-  "Biblical Studies",
-  "Communication",
-  "Communication and Media Studies",
-  "Creative Writing",
-  "Digital Communication and Multimedia",
-  "English",
-  "Foreign Languages and Literature",
-  "French Language and Literature",
-  "Gender Studies",
-  "General Studies",
-  "General Studies and Humanities",
-  "History",
-  "Humanities",
-  "Journalism",
-  "Liberal Arts",
-  "Linguistics",
-  "Mass Communication",
-  "Organizational Communication",
-  "Parks, Recreation and Leisure Facilities Management",
-  "Parks, Recreation and Leisure Studies",
-  "Philosophy",
-  "Public Relations",
-  "Public Relations/Image Management",
-  "Radio and Television",
-  "Religious Studies",
-  "Rhetoric and Composition",
-  "Spanish Language and Literature",
-  "Speech and Rhetoric",
-  "Talmudic Studies",
-  "Theology",
-  "Community Organization and Advocacy",
-  "Geography",
-  "International Relations",
-  "International Studies",
-  "Legal Assistant",
-  "Legal Studies",
-  "Political Science",
-  "Public Administration",
-  "Public Policy Analysis",
-  "Social Work",
-  "Sociology",
-  "Anthropology",
-  "Behavioral Sciences",
-  "Child Development",
-  "Criminology",
-  "Experimental Psychology",
-  "Human Development and Family Studies",
-  "Psychology",
-  "Aerospace Engineering",
-  "Applied Mathematics",
-  "Architecture",
-  "Astrophysics",
-  "Biochemistry",
-  "Biological and Biomedical Sciences",
-  "Biological and Physical Sciences",
-  "Biology",
-  "Biomedical Engineering",
-  "Biomedical Sciences",
-  "Cellular and Molecular Biology",
-  "Chemical Engineering",
-  "Chemistry",
-  "Civil Engineering",
-  "Computer and Information Sciences",
-  "Computer Systems Networking",
-  "Conservation",
-  "Earth Science",
-  "Engineering",
-  "Environmental Science",
-  "Environmental Studies",
-  "Exercise Physiology",
-  "Informatics",
-  "Information Science",
-  "Information Systems Security",
-  "Kinesiology",
-  "Marine Biology",
-  "Mathematics",
-  "Mechanical Engineering",
-  "Microbiology",
-  "Neuroscience",
-  "Petroleum Engineering",
-  "Physics",
-  "Physiology, Pathology, and Related Sciences, Other",
-  "Statistics",
-  "Web Design",
-  "Zoology",
-  "Apparel and Textiles",
-  "Corrections and Criminal Justice",
-  "Criminal Justice",
-  "Electromechanical Technology",
-  "Foods, Nutrition, and Wellness Studies",
-  "Forensic Science and Technology",
-  "Hospitality Administration",
-  "Hotel Management",
-  "Human Sciences (General)",
-  "Human Services",
-  "Law Enforcement Administration",
-  "Police Science",
-  "Sport and Fitness Administration",
-  "One Piece",
-  "Other",
-] as const;
