--- packages/consts/src/shirt-sizes.ts
@@ -1,1 +1,0 @@
-export const SHIRT_SIZES = ["SM", "MD", "LG", "XL", "XXL"] as const;
