--- packages/consts/src/sponsor-tiers.ts
@@ -1,1 +1,0 @@
-export const SPONSOR_TIERS = ["gold", "silver", "bronze", "other"] as const;
