+++ packages/validators/tsconfig.json
@@ -1,8 +1,9 @@
 {
-  "extends": "@knighthacks/tsconfig/base.json",
+  "extends": "@knighthacks/tsconfig/internal-package.json",
   "compilerOptions": {
+    "outDir": "dist",
     "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json"
   },
-  "include": ["*.ts", "src"],
+  "include": ["src"],
   "exclude": ["node_modules"]
 }
